### 2023-11-01 #6004 @danielskatz
@jmlarson1 - thanks for your submission. Has this work been submitted to JOSS in the past? I feel like I remember it, but I might be thinking of it in another context.

Other than this possibility, the submission seems ready to review. Once I hear back from you, I'll work on finding an editor who will select reviewers and run the review process.

Please don't @ tag potential reviewers, as we don't want to clutter anyone's notifications
