+++ README.rst
@@ -63,7 +63,9 @@ Basic Usage
 ===========
 
 Create an ``Ensemble``, then customize with general settings, simulation and generator parameters,
-and an exit condition. Run the following via ``python this_file.py --comms local --nworkers 4``::
+and an exit condition. Run the following via ``python this_file.py --comms local --nworkers 4``:
+
+.. code-block:: python
 
    import numpy as np
 
@@ -74,29 +76,29 @@ and an exit condition. Run the following via ``python this_file.py --comms local
    from libensemble.tools import add_unique_random_streams
 
    if __name__ == "__main__":
-      sampling = Ensemble(parse_args=True)
-      sampling.sim_specs = SimSpecs(
-         sim_f=six_hump_camel,
-         inputs=["x"],
-         outputs=[("f", float)],
-      )
-      sampling.gen_specs = GenSpecs(
-         gen_f=uniform_random_sample,
-         outputs=[("x", float, (2,))],
-         user={
+       sampling = Ensemble(parse_args=True)
+       sampling.sim_specs = SimSpecs(
+           sim_f=six_hump_camel,
+           inputs=["x"],
+           outputs=[("f", float)],
+       )
+       sampling.gen_specs = GenSpecs(
+           gen_f=uniform_random_sample,
+           outputs=[("x", float, (2,))],
+           user={
                "gen_batch_size": 500,
                "lb": np.array([-3, -2]),
                "ub": np.array([3, 2]),
-         },
-      )
+           },
+       )
 
-      sampling.persis_info = add_unique_random_streams({}, sampling.nworkers + 1)
-      sampling.exit_criteria = ExitCriteria(sim_max=101)
-      sampling.run()
-      sampling.save_output(__file__)
+       sampling.persis_info = add_unique_random_streams({}, sampling.nworkers + 1)
+       sampling.exit_criteria = ExitCriteria(sim_max=101)
+       sampling.run()
+       sampling.save_output(__file__)
 
-      if sampling.is_manager:
-         print("Some output data:\n", sampling.H[["x", "f"]][:10])
+       if sampling.is_manager:
+           print("Some output data:\n", sampling.H[["x", "f"]][:10])
 
 Resources
 =========
