### 2023-11-01 #6004 @jbytecode
👋👋👋 Dear  @nickc1,  @jsoishi,  @brian-rose  👋👋👋

**libEnsemble: A complete Python toolkit for dynamic ensembles of calculations**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/6004).

Thank you in advance!
