+++ libensemble/tests/unit_tests/test_executor.py
@@ -431,6 +431,7 @@ def test_procs_and_machinefile_logic():
         task = exctr.submit(calc_type="sim", num_nodes=2, procs_per_node=3, app_args=args_for_sim)
     assert 1
     task = polling_loop(exctr, task, delay=0.05)
+    time.sleep(0.25)
     assert task.finished, "task.finished should be True. Returned " + str(task.finished)
     assert task.state == "FINISHED", "task.state should be FINISHED. Returned " + str(task.state)
 
