### 2023-11-01 5486072
it would be uncommon, but if user sets save_H_on_completion=False and a save_every_k, don't overwrite
