### 2023-11-01 1cc70af
Bump crate-ci/typos from 1.16.21 to 1.16.22
Bumps [crate-ci/typos](https://github.com/crate-ci/typos) from 1.16.21 to 1.16.22.
- [Release notes](https://github.com/crate-ci/typos/releases)
- [Changelog](https://github.com/crate-ci/typos/blob/master/CHANGELOG.md)
- [Commits](https://github.com/crate-ci/typos/compare/v1.16.21...v1.16.22)

---
updated-dependencies:
- dependency-name: crate-ci/typos
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
