+++ .github/workflows/extra.yml
@@ -229,4 +229,4 @@ jobs:
         runs-on: ubuntu-latest
         steps:
         - uses: actions/checkout@v4
-        - uses: crate-ci/typos@v1.16.21
+        - uses: crate-ci/typos@v1.16.22
