### 2023-11-01 9134292
Only return new points to gen when using gen_on_manager
