### 2023-11-01 3fc067c
Support generators as a class
