### 2023-11-02 2b3d99a
Try moving code-sample into README (#1150)
* try moving code-sample into README

* adjust readme formatting method

* try this approach

* maybe the linenos was the problem

* Update README.rst

* give oversubscribed app some time to finish, mostly for macos jobs

* also need it here

---------

Co-authored-by: Jeffrey Larson <jmlarson@anl.gov>
