### 2023-11-02 7b2e780
Feature / Request History options (#1141)
* due to LibeSpecs options being attributes instead of keys, they really shouldn't have double quotes. Plus this makes the tabs resemble the autodoc_pydantic output more. Did this while initially adding the new History options

* move some options, add some whitespace

* simple initial implementations of using H_file_prefix and save_H_with_date in manager

* add _save_H_on_complete, which activates in final_receive_and_kill if save_H_on_completion is set

* fix, plus try options within test_uniform_sampling

* add prefix-option to test

* add output fi
[…]
