+++ docs/data_structures/libE_specs.rst
@@ -191,8 +191,9 @@ libEnsemble is primarily customized by setting options within a ``LibeSpecs`` cl
                 **save_H_and_persis_on_abort** [bool] = ``True``:
                     Save states of ``H`` and ``persis_info`` to file on aborting after an exception.
 
-                **save_H_on_completion** Optional[bool] = ``True``
-                    Save state of ``H`` to file upon completing a workflow.
+                **save_H_on_completion** Optional[bool] = ``False``
+                    Save state of ``H`` to file upon completing a workflow. Also enabled when either ``save_every_k_sims``
+                    or ``save_every_k_gens`` is set.
 
                 **save_H_with_date** Optional[bool] = ``False``
                     Save ``H`` filename contains date and timestamp.
