+++ libensemble/tests/functionality_tests/check_libE_stats.py
@@ -71,9 +71,6 @@ def check_libE_stats(task_datetime=False):
 
 
 if __name__ == "__main__":
-    import warnings
-
-    warnings.filterwarnings("ignore", category=DeprecationWarning)
 
     check_libE_stats()
     check_libE_stats(task_datetime=True)
