+++ libensemble/tests/functionality_tests/check_libE_stats.py
@@ -5,7 +5,11 @@ required. Checks that dates/times are in a valid format.
 
 """
 
+import warnings
+
 from dateutil.parser import parse
+
+warnings.filterwarnings("ignore", category=DeprecationWarning)
 
 infile = "libE_stats.txt"
 
