### 2023-11-03 #6004 @brian-rose
@jbytecode thanks for the invitation, but I don't think I have the right disciplinary expertise in parallel computing to be a good reviewer for this contribution.
