### 2023-11-03 #6004 @jsoishi
@jbytecode I'd be happy to review.
