### 2023-11-03 0bff73f
Prevent duplicate save at end when only save_every_k_gens
