### 2023-11-03 2848379
Trim H save in _save_every_k
