### 2023-11-03 f06c229
Increase timeout in test_procs_and_machinefile_logic
