### 2023-11-04 8ca5a3d
Implement updated `QCommThread`
