+++ libensemble/comms/comms.py
@@ -28,6 +28,7 @@ from abc import ABC, abstractmethod
 
 # from multiprocessing import Process, Queue, Value, Lock
 from multiprocessing import Process, Queue
+from threading import Thread
 from time import time
 from traceback import format_exc
 
@@ -147,6 +148,103 @@ class QComm(Comm):
     def mail_flag(self):
         """Check whether we know a message is ready for receipt."""
         return not self._inbox.empty()
+
+
+class QCommThread(Comm):
+    """Launch a user function in a thread with an attached QComm."""
+
+    def __init__(self, main, nworkers, *args, **kwargs):
+        self.inbox = thread_queue.Queue()
+        self.outbox = thread_queue.Queue()
+        self._result = None
+        self._exception = None
+        self._done = False
+        comm = QComm(self.inbox, self.outbox, nworkers)
+        self.thread = Thread(target=QCommThread._qcomm_main, args=(comm, main) + args, kwargs=kwargs)
+
+    def _is_result_msg(self, msg):
+        """Return true if message indicates final result (and set result/except)."""
+        if len(msg) and isinstance(msg[0], CommResult):
+            self._result = msg[0].value
+            self._done = True
+            return True
+        if len(msg) and isinstance(msg[0], CommResultErr):
+            self._exception = msg[0]
+            self._done = True
+            return True
+        return False
+
+    def send(self, *args):
+        """Send a message to the thread (called from creator)"""
+        self.inbox.put(args)
+
+    def recv(self, timeout=None):
+        """Return a message from the thread or raise TimeoutError."""
+        try:
+            if self._done:
+                raise CommFinishedException()
+            if not self.outbox.empty():
+                msg = self.outbox.get()
+            else:
+                msg = self.outbox.get(timeout=timeout)
+            if self._is_result_msg(msg):
+                raise CommFinishedException()
+            return msg
+        except thread_queue.Empty:
+            raise Timeout()
+
+    def mail_flag(self):
+        """Check whether we know a message is ready for receipt."""
+        return not self.outbox.empty()
+
+    def run(self):
+        """Start the thread."""
+        self.thread.start()
+
+    def result(self, timeout=None):
+        """Join and return the thread main result (or re-raise an exception)."""
+        get_timeout = _timeout_fun(timeout)
+        while not self._done and (timeout is None or timeout >= 0):
+            try:
+                msg = self.outbox.get(timeout=timeout)
+            except thread_queue.Empty:
+                raise Timeout()
+            self._is_result_msg(msg)
+            timeout = get_timeout()
+        if not self._done:
+            raise Timeout()
+        self.thread.join(timeout=timeout)
+        if self.running:
+            raise Timeout()
+        if self._exception is not None:
+            raise RemoteException(self._exception.msg, self._exception.exc)
+        return self._result
+
+    def terminate(self, timeout=None):
+        """Terminate the thread."""
+        raise Timeout()
+
+    @property
+    def running(self):
+        """Check if the thread is running."""
+        return self.thread.is_alive()
+
+    @staticmethod
+    def _qcomm_main(comm, main, *args, **kwargs):
+        """Main routine -- handles return values and exceptions."""
+        try:
+            _result = main(comm, *args, **kwargs)
+            comm.send(CommResult(_result))
+        except Exception as e:
+            comm.send(CommResultErr(str(e), format_exc()))
+            raise e
+
+    def __enter__(self):
+        self.run()
+        return self
+
+    def __exit__(self, etype, value, traceback):
+        self.thread.join()
 
 
 class QCommProcess(Comm):
