### 2023-11-04 6f4aeb6
Add `local_comms` to `LibeSpecs`
