### 2023-11-04 3635052
Enable threading in `libE_local`
