+++ libensemble/specs.py
@@ -206,10 +206,7 @@ class LibeSpecs(BaseModel):
     """
 
     comms: Optional[str] = "mpi"
-    """ Manager/Worker communications mode. ``'mpi'``, ``'local'``, or ``'tcp'`` """
-
-    local_comms: Optional[str] = "multiprocessing"
-    """ Local Manager/Worker communications mode. ``'multiprocessing'`` or ``'threading'`` """
+    """ Manager/Worker communications mode. ``'mpi'``, ``'local'``, ``'local_threading'``, or ``'tcp'`` """
 
     nworkers: Optional[int]
     """ Number of worker processes in ``"local"`` or ``"tcp"``."""
@@ -503,12 +500,7 @@ class LibeSpecs(BaseModel):
 
     @validator("comms")
     def check_valid_comms_type(cls, value):
-        assert value in ["mpi", "local", "tcp"], "Invalid comms type"
-        return value
-    
-    @validator("local_comms")
-    def check_valid_local_comms_type(cls, value):
-        assert value in ["multiprocessing", "threading"], "Invalid local comms type"
+        assert value in ["mpi", "local", "local_threading", "tcp"], "Invalid comms type"
         return value
 
     @validator("platform_specs")
