### 2023-11-06 1314d56
Call finalize if function exists
