+++ libensemble/run_calculation.py
@@ -24,6 +24,7 @@ from libensemble.message_numbers import (
     CALC_EXCEPTION,
     EVAL_GEN_TAG,
     EVAL_SIM_TAG,
+    EVAL_FINAL_GEN_TAG,
     MAN_SIGNAL_FINISH,
     MAN_SIGNAL_KILL,
     PERSIS_STOP,
@@ -198,6 +199,7 @@ class RunCalc:
                 calc_id = str(Work["libE_info"]["gen_count"])
             else:
                 calc_id = str(self.calc_iter[calc_type])
+
         # Add a right adjust (minimum width).
         calc_id = calc_id.rjust(5, " ")
 
@@ -205,7 +207,13 @@ class RunCalc:
 
         try:
             logger.debug(f"Starting {enum_desc}: {calc_id}")
-            calc = self._run_calc[calc_type]
+
+            # SH TODO May use a different calc_type
+            if Work["libE_info"].get("finalize", False):
+                calc = self._run_calc[EVAL_FINAL_GEN_TAG]
+            else:
+                calc = self._run_calc[calc_type]
+
             with timer:  # check works when making gen dirs (and when there is a workflow dir). Matrix of checks
                 if self.EnsembleDirectory.use_calc_dirs(calc_type):
                     loc_stack, calc_dir = self.EnsembleDirectory.prep_calc_dir(
