### 2023-11-06 c25e6e0
enable workflow dispatch, temporary push: directive to register complete-CI workflow
