### 2023-11-06 cfdcc09
Manual extra-trigger for complete-CI on develop (#1157)
* enable workflow dispatch, temporary push: directive to register complete-CI workflow

* remove temporary push:
