### 2023-11-06 e8f20db
tentative change to support both pydantic 1 and pydantic 2
