+++ libensemble/specsV1.py
@@ -1,0 +1,601 @@
+import random
+import secrets
+from pathlib import Path
+from typing import Any, Callable, List, Optional, Tuple, Union
+
+import numpy as np
+from pydantic import BaseConfig, BaseModel, Field, root_validator, validator
+
+from libensemble.alloc_funcs.give_sim_work_first import give_sim_work_first
+from libensemble.gen_funcs.sampling import latin_hypercube_sample
+from libensemble.resources.platforms import Platform
+from libensemble.sim_funcs.one_d_func import one_d_example
+from libensemble.utils.specs_checkers import (
+    _check_any_workers_and_disable_rm_if_tcp,
+    _check_exit_criteria,
+    _check_H0,
+    _check_output_fields,
+)
+
+_UNRECOGNIZED_ERR = "Unrecognized field. Check closely for typos, or libEnsemble's docs"
+_OUT_DTYPE_ERR = "Unable to coerce '{}' into a NumPy dtype. It should be a list of 2-tuples or 3-tuples"
+_IN_INVALID_ERR = "Value should be a list of field names (a list of strings)"
+_UFUNC_INVALID_ERR = "Specified sim_f or gen_f is not callable. It should be a user function"
+
+BaseConfig.arbitrary_types_allowed = True
+BaseConfig.allow_population_by_field_name = True
+BaseConfig.extra = "forbid"
+BaseConfig.error_msg_templates = {
+    "value_error.extra": _UNRECOGNIZED_ERR,
+    "type_error.callable": _UFUNC_INVALID_ERR,
+}
+BaseConfig.validate_assignment = True
+
+__all__ = ["SimSpecs", "GenSpecs", "AllocSpecs", "ExitCriteria", "LibeSpecs", "_EnsembleSpecs"]
+
+
+class SimSpecs(BaseModel):
+    """
+    Specifications for configuring a Simulation Function.
+    """
+
+    sim_f: Callable = one_d_example
+    """
+    Python function matching the ``sim_f`` interface. Evaluates parameters
+    produced by a generator function.
+    """
+
+    inputs: List[str] = Field([], alias="in")
+    """
+    List of **field names** out of the complete history to pass
+    into the simulation function upon calling.
+    """
+
+    persis_in: Optional[List[str]] = []
+    """
+    List of **field names** to send to a persistent simulation function
+    throughout the run, following initialization.
+    """
+
+    # list of tuples for dtype construction
+    outputs: List[Union[Tuple[str, Any], Tuple[str, Any, Union[int, Tuple]]]] = Field([], alias="out")
+    """
+    List of 2- or 3-tuples corresponding to NumPy dtypes.
+    e.g. ``("dim", int, (3,))``, or ``("path", str)``.
+    Typically used to initialize an output array within the simulation function:
+    ``out = np.zeros(100, dtype=sim_specs["out"])``.
+    Also necessary to construct libEnsemble's history array.
+    """
+
+    globus_compute_endpoint: Optional[str] = ""
+    """
+    A Globus Compute (https://www.globus.org/compute) ID corresponding to an active endpoint on a remote system.
+    libEnsemble's workers will submit simulator function instances to this endpoint instead of
+    calling them locally.
+    """
+
+    user: Optional[dict] = {}
+    """
+    A user-data dictionary to place bounds, constants, settings, or other parameters for customizing
+    the simulator function.
+    """
+
+    @validator("outputs", pre=True)
+    def check_valid_out(cls, v):
+        try:
+            _ = np.dtype(v)
+        except TypeError:
+            raise ValueError(_OUT_DTYPE_ERR.format(v))
+        else:
+            return v
+
+    @validator("inputs", "persis_in", pre=True)
[…]
