### 2023-11-06 0b365f8
Bump crate-ci/typos from 1.16.22 to 1.16.23
Bumps [crate-ci/typos](https://github.com/crate-ci/typos) from 1.16.22 to 1.16.23.
- [Release notes](https://github.com/crate-ci/typos/releases)
- [Changelog](https://github.com/crate-ci/typos/blob/master/CHANGELOG.md)
- [Commits](https://github.com/crate-ci/typos/compare/v1.16.22...v1.16.23)

---
updated-dependencies:
- dependency-name: crate-ci/typos
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
