+++ .github/workflows/extra.yml
@@ -216,4 +216,4 @@ jobs:
         runs-on: ubuntu-latest
         steps:
         - uses: actions/checkout@v4
-        - uses: crate-ci/typos@v1.16.22
+        - uses: crate-ci/typos@v1.16.23
