+++ .github/workflows/basic.yml
@@ -140,4 +140,4 @@ jobs:
         runs-on: ubuntu-latest
         steps:
         - uses: actions/checkout@v4
-        - uses: crate-ci/typos@v1.16.22
+        - uses: crate-ci/typos@v1.16.23
