### 2023-11-06 7439645
Add comment to `terminate`
