### 2023-11-06 e86fc54
attempts to set specs and platforms modules based on pydantic version
