### 2023-11-07 #6031 @editorialbot
[REVIEW]: libEnsemble: A complete Python toolkit for dynamic ensembles of calculations
**Submitting author:** @jmlarson1 (<a href="http://orcid.org/0000-0001-9924-2082">Jeffrey Larson</a>)
**Repository:** https://github.com/Libensemble/libensemble
**Branch with paper.md** (empty if default branch): joss-develop
**Version:** v1.1.0
