### 2023-11-07 9833c9d
better variables for determining pydantic version, and specs_dump function for dumping specs based on pydantic version
