+++ libensemble/resources/platforms/__init__.py
@@ -1,8 +1,8 @@
-import pydantic
+from libensemble.utils.pydantic_support import pydanticV1, pydanticV2
 
-if pydantic.__version__[0] == "1":
+if pydanticV1:
     from .platformsV1 import Platform  # noqa: F401
     from .platformsV1 import get_platform  # noqa: F401
-elif pydantic.__version__[0] == "2":
+elif pydanticV2:
     from .platformsV2 import Platform  # noqa: F401
     from .platformsV2 import get_platform  # noqa: F401
