+++ libensemble/specs/__init__.py
@@ -1,6 +1,6 @@
-import pydantic
+from libensemble.utils.pydantic_support import pydanticV1, pydanticV2
 
-if pydantic.__version__[0] == "1":
+if pydanticV1:
     from .specsV1 import AllocSpecs, ExitCriteria, GenSpecs, LibeSpecs, SimSpecs, _EnsembleSpecs  # noqa: F401
-elif pydantic.__version__[0] == "2":
+elif pydanticV2:
     from .specsV2 import AllocSpecs, ExitCriteria, GenSpecs, LibeSpecs, SimSpecs, _EnsembleSpecs  # noqa: F401
