+++ libensemble/ensemble.py
@@ -4,7 +4,6 @@ import logging
 from typing import Optional
 
 import numpy.typing as npt
-import pydantic
 import tomli
 import yaml
 
@@ -15,6 +14,7 @@ from libensemble.specs import AllocSpecs, ExitCriteria, GenSpecs, LibeSpecs, Sim
 from libensemble.tools import add_unique_random_streams
 from libensemble.tools import parse_args as parse_args_f
 from libensemble.tools import save_libE_output
+from libensemble.utils.pydantic_support import specs_dump
 
 ATTR_ERR_MSG = 'Unable to load "{}". Is the function or submodule correctly named?'
 ATTR_ERR_MSG = "\n" + 10 * "*" + ATTR_ERR_MSG + 10 * "*" + "\n"
@@ -323,10 +323,7 @@ class Ensemble:
 
         # Cast new libE_specs temporarily to dict
         if not isinstance(new_specs, dict):
-            if pydantic.__version__[0] == "2":
-                new_specs = new_specs.model_dump(by_alias=True, exclude_none=True, exclude_unset=True)
-            elif pydantic.__version__[0] == "1":
-                new_specs = new_specs.dict(by_alias=True, exclude_none=True, exclude_unset=True)
+            new_specs = specs_dump(new_specs, by_alias=True, exclude_none=True, exclude_unset=True)
 
         # Unset "comms" if we already have a libE_specs that contains that field, that came from parse_args
         if new_specs.get("comms") and hasattr(self._libE_specs, "comms") and self.parsed:
