+++ README.rst
@@ -25,13 +25,13 @@
    :target: https://libensemble.readthedocs.org/en/latest/
    :alt: Documentation Status
 
-.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
-   :target: https://github.com/psf/black
-   :alt: Code style: black
-
 .. image:: https://joss.theoj.org/papers/8d3b1a704c83affe55284dd00857bdb0/status.svg
    :target: https://joss.theoj.org/papers/8d3b1a704c83affe55284dd00857bdb0
    :alt: JOSS Status
+
+.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
+   :target: https://github.com/psf/black
+   :alt: Code style: black
 
 |
 
