+++ README.rst
@@ -33,6 +33,10 @@
    :target: https://github.com/psf/black
    :alt: Code style: black
 
+.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
+   :target: https://github.com/psf/black
+   :alt: Code style: black
+
 |
 
 .. after_badges_rst_tag
