### 2023-11-08 f859b25
Add release notes for v1.1.0
