### 2023-11-08 f87bfd9
revoke multitask_gp_ax test for python 3.12, plus ax-platform dep install on that python
