+++ .github/workflows/extra.yml
@@ -111,7 +111,7 @@ jobs:
             pip install --upgrade git+https://github.com/bandframework/surmise.git@develop
 
         - name: Build ytopt and dependencies
-          if: matrix.python-version != '3.11' && matrix.os != 'macos-latest'
+          if: matrix.python-version <= '3.10' && matrix.os != 'macos-latest'
           run: |
             pip install scikit-learn==1.1.2
             pip install pandas==1.5.3
@@ -187,8 +187,8 @@ jobs:
             pip install -e .
             flake8 libensemble
 
-        - name: Remove ytopt-heffte test on Python 3.11 (easy way)
-          if: matrix.python-version == '3.11'
+        - name: Remove ytopt-heffte test on Python 3.11 and 3.12
+          if: matrix.python-version <= '3.11'
           run: |
             rm ./libensemble/tests/regression_tests/test_ytopt_heffte.py
 
