### 2023-11-08 d116280
adjust various steps to not run installs/tests on newer python versions
