+++ .github/workflows/extra.yml
@@ -111,7 +111,7 @@ jobs:
             pip install --upgrade git+https://github.com/bandframework/surmise.git@develop
 
         - name: Build ytopt and dependencies
-          if: matrix.python-version != '3.11' && matrix.os != 'macos-latest'
+          if: (matrix.python-version != '3.12' || matrix.python-version != '3.11') && matrix.os != 'macos-latest'
           run: |
             pip install scikit-learn==1.1.2
             pip install pandas==1.5.3
@@ -131,7 +131,7 @@ jobs:
             cd ..
 
         - name: Install generator dependencies for Ubuntu tests
-          if: matrix.os == 'ubuntu-latest' && steps.cache.outputs.cache-hit != 'true'
+          if: matrix.os == 'ubuntu-latest' && steps.cache.outputs.cache-hit != 'true' && matrix.python-version != '3.12'
           run: |
             conda install octave
             sudo apt-get install bc
@@ -187,14 +187,11 @@ jobs:
             pip install -e .
             flake8 libensemble
 
-        - name: Remove ytopt-heffte test on Python 3.11 (easy way)
-          if: matrix.python-version == '3.11'
+        - name: Remove various tests on newer pythons
+          if: matrix.python-version == '3.11' || matrix.python-version == '3.12'
           run: |
             rm ./libensemble/tests/regression_tests/test_ytopt_heffte.py
-
-        - name: Remove gp_multitask_ax test on Python 3.12
-          if: matrix.python-version == '3.12'
-          run: |
+            rm ./libensemble/tests/regression_tests/test_persistent_gp.py
             rm ./libensemble/tests/regression_tests/test_persistent_gp_multitask_ax.py
 
         - name: Run extensive tests, Ubuntu
