+++ libensemble/tests/unit_tests/test_executor_balsam.py
@@ -10,9 +10,10 @@ from dataclasses import dataclass
 import mock
 import pytest
 
-from libensemble.executors.executor import Application, Executor, ExecutorException, TimeoutExpired
-
 warnings.filterwarnings("ignore", category=DeprecationWarning)
+
+
+from libensemble.executors.executor import Application, Executor, ExecutorException, TimeoutExpired  # noqa E402
 
 
 def setup_module(module):
