### 2023-11-08 b64de7c
removing error-on-warnings in extensive tests, for now
