### 2023-11-08 84326dd
always install octave and tasmanian
