### 2023-11-08 7148a26
ensure that if we did a temporary cast of a new libE_specs to dict, it gets cast back to a class upon setting. Passthrough nworkers attribute upon setting libE_specs if the Ensemble doesn't have it, but libE_specs does
