### 2023-11-08 19dfba5
Testing/revoke torch for 312 (#1161)
* Removing badge

* revoke multitask_gp_ax test for python 3.12, plus ax-platform dep install on that python

* Removing ytopt from 3.12 as well

* adjust various steps to not run installs/tests on newer python versions

* fix heffte exe move

* try ignoring deprecationwarning for balsam test

* need to filter warning before Executors imported

* removing error-on-warnings in extensive tests, for now

* always install octave and tasmanian

* revoke test_heffte

---------

Co-authored-by: Jeffrey Larson <jmlarson@anl.gov>
