### 2023-11-08 6199ef2
Some symmetry in changelog wording
