+++ docs/welcome.rst
@@ -14,7 +14,7 @@
 
   |
 
-  .. image:: https://github.com/Libensemble/libensemble/actions/workflows/basic.yml/badge.svg?branch=develop
+  .. image:: https://github.com/Libensemble/libensemble/actions/workflows/extra.yml/badge.svg?branch=main
     :target: https://github.com/Libensemble/libensemble/actions
 
   .. image:: https://codecov.io/github/Libensemble/libensemble/graph/badge.svg
