### 2023-11-10 c7e13b2
Moving exception def
