### 2023-11-12 Libensemble/libensemble#1167 @SergeyYakubov
Problems running "Simple introduction tutorial"
trying to run the tutorial from https://libensemble.readthedocs.io/en/main/tutorials/local_sine_tutorial.html,
I get errors like

```
from libensemble import Ensemble, LibeSpecs, SimSpecs, GenSpecs, ExitCriteria
ImportError: cannot import name 'LibeSpecs' from 'libensemble' (/Library/Frameworks/Python.framework/Versions/3.10/lib/python3.10/site-packages/libensemble/__init__.py)

```
fixing error with imports (use `from libensemble.specs import LibeSpecs, SimSpecs, GenSpecs,  ExitCriteria
`)

I got another error:

```
File "/Users/35y/projects/reviews/libEnsemble/calling_script.py", line 31, in <module>
    ensemble.add_random_streams()  # setup the random streams unique to each worker
  File "/Library/Frameworks/Python.framework/Versions/3.10/lib/python3.10/site-packages/libensemble/ensemble.py", line 535, in add_random_streams
    nstreams = self._nworkers()
  File "/Library/Frameworks/Python.framework/Versions/3.10/lib/python3.10/site-packages/libensemble/ensemble.py", line 396, in _nworkers
    return self.libE_specs.nworkers
AttributeError: 'ExitCriteria' object has no attribute 'nworkers'

```

so, I guess the tutorial has to be updated for the latest libEnsemble version.
