### 2023-11-12 #6031 @SergeyYakubov
Hi @jmlarson1, I started looking into your paper.

Looks pretty interesting. Couple of issues as of now:

1.  In the "statement of need," I do not see a clear description of the "target audience" as required by the checklist. Could you please add this to the paper (and maybe to the docs asa well)?
2. Cannot run simple tutorial  - see https://github.com/Libensemble/libensemble/issues/1167
