### 2023-11-12 df57ba6
fix Ensemble instantiation and lines beneath if name == main in intro tutorial
