### 2023-11-12 Libensemble/libensemble#1167 @jlnav
Hi @SergeyYakubov ,

Whoops, sorry about that! Thank you for opening this issue to point that out, we're fixing those tutorial bugs and any others that may show up in #1168 .

Please let us know if there's anything else you need!
