### 2023-11-13 ae33c25
ensure filenames and import namespaces are correct. update example files to match modern tutorial
