+++ examples/tutorials/simple_sine/tutorial_gen.py
@@ -1,13 +1,11 @@
 import numpy as np
 
 
-def gen_random_sample(H, persis_info, gen_specs, _):
-    # underscore parameter for internal/testing arguments
-
-    # Pull out user parameters to perform calculations
+def gen_random_sample(Input, persis_info, gen_specs):
+    # Pull out user parameters
     user_specs = gen_specs["user"]
 
-    # Get lower and upper bounds from gen_specs
+    # Get lower and upper bounds
     lower = user_specs["lower"]
     upper = user_specs["upper"]
 
@@ -15,11 +13,11 @@ def gen_random_sample(H, persis_info, gen_specs, _):
     num = len(lower)
     batch_size = user_specs["gen_batch_size"]
 
-    # Create array of 'batch_size' zeros
-    out = np.zeros(batch_size, dtype=gen_specs["out"])
+    # Create empty array of "batch_size" zeros. Array dtype should match "out" fields
+    Output = np.zeros(batch_size, dtype=gen_specs["out"])
 
-    # Replace those zeros with the random numbers
-    out["x"] = persis_info["rand_stream"].uniform(lower, upper, (batch_size, num))
+    # Set the "x" output field to contain random numbers, using random stream
+    Output["x"] = persis_info["rand_stream"].uniform(lower, upper, (batch_size, num))
 
     # Send back our output and persis_info
-    return out, persis_info
+    return Output, persis_info
