### 2023-11-13 88f37e2
remove "tutorial_" prefix from simple sine tutorial files
