+++ examples/tutorials/simple_sine/calling.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
-from tutorial_gen import gen_random_sample
-from tutorial_sim import sim_find_sine
+from gen import gen_random_sample
+from sim import sim_find_sine
 
 from libensemble import Ensemble
 from libensemble.specs import ExitCriteria, GenSpecs, LibeSpecs, SimSpecs
