+++ docs/tutorials/local_sine_tutorial.rst
@@ -64,11 +64,11 @@ need to write a new allocation function.
 
         Later on, we'll populate :class:`gen_specs<libensemble.specs.GenSpecs>` and ``persis_info`` when we initialize libEnsemble.
 
-        For now, create a new Python file named ``tutorial_gen.py``. Write the following:
+        For now, create a new Python file named ``gen.py``. Write the following:
 
         .. code-block:: python
             :linenos:
-            :caption: examples/tutorials/simple_sine/tutorial_gen.py
+            :caption: examples/tutorials/simple_sine/gen.py
 
             import numpy as np
 
@@ -131,11 +131,11 @@ need to write a new allocation function.
         The only new parameter here is :ref:`sim_specs<datastruct-sim-specs>`, which
         serves a purpose similar to the :class:`gen_specs<libensemble.specs.GenSpecs>` dictionary.
 
-        Create a new Python file named ``tutorial_sim.py``. Write the following:
+        Create a new Python file named ``sim.py``. Write the following:
 
         .. code-block:: python
             :linenos:
-            :caption: examples/tutorials/simple_sine/tutorial_sim.py
+            :caption: examples/tutorials/simple_sine/sim.py
 
             import numpy as np
 
@@ -193,8 +193,8 @@ need to write a new allocation function.
             import numpy as np
             from libensemble import Ensemble
             from libensemble.specs import LibeSpecs, SimSpecs, GenSpecs, ExitCriteria
-            from tutorial_gen import gen_random_sample
-            from tutorial_sim import sim_find_sine
+            from gen import gen_random_sample
+            from sim import sim_find_sine
 
             libE_specs = LibeSpecs(nworkers=4, comms="local")
 
@@ -327,8 +327,8 @@ need to write a new allocation function.
                 import numpy as np
                 from libensemble import Ensemble
                 from libensemble.specs import LibeSpecs, SimSpecs, GenSpecs, ExitCriteria
-                from tutorial_gen import gen_random_sample
-                from tutorial_sim import sim_find_sine
+                from gen import gen_random_sample
+                from sim import sim_find_sine
 
                 libE_specs = LibeSpecs(nworkers=4, comms="local")
 
