+++ .github/workflows/basic.yml
@@ -96,7 +96,7 @@ jobs:
             pip install mpmath
             conda install numpy nlopt scipy
 
-            git clone -b refactor/pounders_API https://github.com/POptUS/IBCDFO.git
+            git clone --recurse-submodules -b refactor/pounders_API https://github.com/POptUS/IBCDFO.git
             pushd IBCDFO/minq/py/minq5/
             export PYTHONPATH="$PYTHONPATH:$(pwd)"
             popd
