+++ .github/workflows/basic.yml
@@ -88,6 +88,7 @@ jobs:
             git clone --recurse-submodules -b refactor/pounders_API https://github.com/POptUS/IBCDFO.git
             pushd IBCDFO/minq/py/minq5/
             export PYTHONPATH="$PYTHONPATH:$(pwd)"
+            echo "PYTHONPATH=$PYTHONPATH" >> $GITHUB_ENV
             popd
             pushd IBCDFO/ibcdfo_pypkg/
             pip install -e .
