### 2023-11-13 207b460
Better error message; adding mpi
