### 2023-11-14 09a078c
Default to local unless MPI detected
