### 2023-11-14 878fd55
Make parse_args default to local if nworkers option set
