### 2023-11-14 34ea71d
Add currently failing test with parse_args False and nworkers set
