### 2023-11-14 fe6648b
Removing extra windows condition
