+++ .github/workflows/basic.yml
@@ -99,7 +99,7 @@ jobs:
             conda install mpi4py ${{ matrix.mpi-version }}
 
         - name: Install generator dependencies
-          if: matrix.os != 'windows-latest' && steps.cache.outputs.cache-hit != 'true'
+          if: steps.cache.outputs.cache-hit != 'true'
           run: |
             python -m pip install --upgrade pip
             pip install mpmath
