### 2023-11-14 ac5f3f3
remove $ from shell lines, add another History doc reference to H_in line
