### 2023-11-14 73f2ea9
emphasize lines, and ensure more .run()s are underneath if name == main
