### 2023-11-14 3a634d1
renaming Input and Output to InputArray and OutputArray, but another solutions may be better
