### 2023-11-14 5c76d95
undo default functions, but enforce they've been set by runtime. then some logic for determining sim/gen specs are correct. But are the dict specs getting cast? What about trailing commas?
