### 2023-11-15 ade68f1
Updates to sine tutorial
