+++ examples/tutorials/simple_sine/sim.py
@@ -1,12 +1,12 @@
 import numpy as np
 
 
-def sim_find_sine(Input, _, sim_specs):
+def sim_find_sine(InputArray, _, sim_specs):
     # Create an output array of a single zero
-    Output = np.zeros(1, dtype=sim_specs["out"])
+    OutputArray = np.zeros(1, dtype=sim_specs["out"])
 
-    # Set the zero to the sine of the Input value
-    Output["y"] = np.sin(Input["x"])
+    # Set the zero to the sine of the InputArray value
+    OutputArray["y"] = np.sin(InputArray["x"])
 
     # Send back our output
-    return Output
+    return OutputArray
