### 2023-11-15 ce33e54
add pydantic validate_arguments decorator to libE, small default adjustments, add unit test to test avoided validation error from before
