+++ libensemble/tests/functionality_tests/test_persistent_uniform_sampling.py
@@ -46,8 +46,6 @@ if __name__ == "__main__":
         "out": [("f", float), ("grad", float, n)],
     }
 
-    sim_specs["in"] = (["x", "obj_component"],)
-
     gen_specs = {
         "persis_in": ["x", "f", "grad", "sim_id"],
         "out": [("x", float, (n,))],
