+++ libensemble/specs.py
@@ -569,7 +569,7 @@ class _EnsembleSpecs(BaseModel):
     persis_info: Optional[dict]
     """ Per-worker information and structures to be passed between user function instances. """
 
-    alloc_specs: Optional[AllocSpecs]
+    alloc_specs: Optional[AllocSpecs] = AllocSpecs()
     """ Specifications for the allocation function. """
 
     nworkers: Optional[int]
@@ -582,12 +582,12 @@ class _EnsembleSpecs(BaseModel):
     def check_provided_ufuncs(cls, values):
         sim_specs = values.get("sim_specs")
         assert hasattr(sim_specs, "sim_f"), "Simulation function not provided to SimSpecs."
-        assert isinstance(sim_specs.sim_f, Callable), "Provided simulation function is not callable."
+        assert isinstance(sim_specs.sim_f, Callable), "Simulation function is not callable."
 
         if values.get("alloc_specs").alloc_f.__name__ != "give_pregenerated_sim_work":
             gen_specs = values.get("gen_specs")
             assert hasattr(gen_specs, "gen_f"), "Generator function not provided to GenSpecs."
-            assert isinstance(gen_specs.gen_f, Callable), "Provided Generator function is not callable."
+            assert isinstance(gen_specs.gen_f, Callable), "Generator function is not callable."
 
         return values
 
