### 2023-11-15 3fc5d46
fixes to MPI calling scripts in simple tutorial, and corresponding example calling script
