### 2023-11-15 7f5507f
Put back matplotlib import in MPI
