### 2023-11-15 980d9da
Remove gen and sim exercises
