+++ examples/tutorials/simple_sine/calling.py
@@ -6,30 +6,29 @@ from sim import sim_find_sine
 from libensemble import Ensemble
 from libensemble.specs import ExitCriteria, GenSpecs, LibeSpecs, SimSpecs
 
-libE_specs = LibeSpecs(nworkers=4, comms="local")
-
-gen_specs = GenSpecs(
-    gen_f=gen_random_sample,  # Our generator function
-    out=[("x", float, (1,))],  # gen_f output (name, type, size)
-    user={
-        "lower": np.array([-3]),  # lower boundary for random sampling
-        "upper": np.array([3]),  # upper boundary for random sampling
-        "gen_batch_size": 5,  # number of x's gen_f generates per call
-    },
-)
-
-sim_specs = SimSpecs(
-    sim_f=sim_find_sine,  # Our simulator function
-    inputs=["x"],  # Input field names. "x" from gen_f output
-    out=[("y", float)],  # sim_f output. "y" = sine("x")
-)
-
-exit_criteria = ExitCriteria(sim_max=80)  # Stop libEnsemble after 80 simulations
-
-ensemble = Ensemble(sim_specs, gen_specs, exit_criteria, libE_specs)
-ensemble.add_random_streams()  # setup the random streams unique to each worker
-
 if __name__ == "__main__":  # Python-quirk required on macOS and windows
+    libE_specs = LibeSpecs(nworkers=4, comms="local")
+
+    gen_specs = GenSpecs(
+        gen_f=gen_random_sample,  # Our generator function
+        out=[("x", float, (1,))],  # gen_f output (name, type, size)
+        user={
+            "lower": np.array([-3]),  # lower boundary for random sampling
+            "upper": np.array([3]),  # upper boundary for random sampling
+            "gen_batch_size": 5,  # number of x's gen_f generates per call
+        },
+    )
+
+    sim_specs = SimSpecs(
+        sim_f=sim_find_sine,  # Our simulator function
+        inputs=["x"],  # Input field names. "x" from gen_f output
+        out=[("y", float)],  # sim_f output. "y" = sine("x")
+    )
+
+    exit_criteria = ExitCriteria(sim_max=80)  # Stop libEnsemble after 80 simulations
+
+    ensemble = Ensemble(sim_specs, gen_specs, exit_criteria, libE_specs)
+    ensemble.add_random_streams()  # setup the random streams unique to each worker
     ensemble.run()  # start the ensemble. Blocks until completion.
 
     history = ensemble.H  # start visualizing our results
