### 2023-11-15 70b1bf8
Indent matplotlib code
