### 2023-11-16 Libensemble/libensemble#1167 @shuds13
@SergeyYakubov

The tutorial has been fixed. Everything should work now.

Thanks for finding this.
