### 2023-11-16 ebded8b
Update autodoc-pydantic requirement from <2 to <3
Updates the requirements on [autodoc-pydantic](https://github.com/mansenfranzen/autodoc_pydantic) to permit the latest version.
- [Changelog](https://github.com/mansenfranzen/autodoc_pydantic/blob/main/changelog.rst)
- [Commits](https://github.com/mansenfranzen/autodoc_pydantic/compare/v0.1.1...v2.0.1)

---
updated-dependencies:
- dependency-name: autodoc-pydantic
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
