### 2023-11-16 #6031 @shuds13
Hi @SergeyYakubov, we have fixed the tutorial. I have also updated our [local copy of the paper](https://github.com/Libensemble/libensemble/blob/joss-develop/docs/papers/joss/paper.md?plain=1#L59)  to provide a target audience. The addition is quite brief, but I think as we describe example use cases above, this may be sufficient. Let me know if you think it would be useful to add more clarification.
