### 2023-11-16 c9d7520
Reorganizing rosenbrock.py logic
