### 2023-11-16 5b206e9
Reorganizing rosenbrock.py logic
