### 2023-11-16 4b37060
Removing some old APOSMM-specific code
