+++ libensemble/alloc_funcs/fast_alloc_and_pausing.py
@@ -124,14 +124,6 @@ def give_sim_work_first(W, H, sim_specs, gen_specs, alloc_specs, persis_info, li
                     np.logical_or(H["sim_ended"][last_size:], H["paused"][last_size:])
                 ):
                     break
-                # Don't call APOSMM if there are runs going but none need advancing
-                if len(persis_info[lw]["run_order"]):
-                    runs_needing_to_advance = np.zeros(len(persis_info[lw]["run_order"]), dtype=bool)
-                    for run, inds in enumerate(persis_info[lw]["run_order"].values()):
-                        runs_needing_to_advance[run] = np.all(H["sim_ended"][inds])
-
-                    if not np.any(runs_needing_to_advance):
-                        break
 
             # Give gen work
             i = idle_workers[0]
