### 2023-11-16 ab84e4f
Covering const usage in sim_func
