### 2023-11-16 eafb315
Trying to debug borehole kills
