### 2023-11-17 82e4745
Trying to debug borehole kills
