### 2023-11-17 cacd1a9
Trying to debug borehole kills
