### 2023-11-17 a93252c
Fix up detect_comms_env
