### 2023-11-20 1f1c4e0
fixes and improvements for libE_specs setting and nworkers attributes passthrough, plus a test
