### 2023-11-20 b68d013
huge refactoring. only a single specs.py for now. validators.py contains validators for both pydantic versions. pydantic_bindings.py binds validators and other configs to their corresponding classes. only one specs_checkers, utility functions to get/set attributes of pydantic objects
