### 2023-11-21 526b508
refactoring docs, turning both platforms.py files back into a single file
