+++ libensemble/resources/platforms/platformsV1.py
@@ -1,353 +1,0 @@
-"""Module for platform specification
-
-This module defines the Platform class which can be used to determine a platform
-(computing system) attributes. Many known systems are provided.
-
-These can be specified by the libE_specs options ``platform_specs`` (recommended).
-It may also be specified, for known systems, via a string in the ``platform``
-option or the environment variable ``LIBE_PLATFORM``.
-"""
-
-import os
-import subprocess
-from typing import Optional
-
-from pydantic import BaseConfig, BaseModel, root_validator, validator
-
-BaseConfig.validate_assignment = True
-
-
-class PlatformException(Exception):
-    """Platform module exception"""
-
-
-class Platform(BaseModel):
-    """
-    Class to define attributes of a target platform.
-
-    All are optional, and any not defined will be determined by libEnsemble's auto-detection.
-    """
-
-    mpi_runner: Optional[str]
-    """MPI runner: One of ``"mpich"``, ``"openmpi"``, ``"aprun"``,
-    ``"srun"``, ``"jsrun"``, ``"msmpi"``, ``"custom"`` """
-
-    runner_name: Optional[str]
-    """Literal string of MPI runner command. Only needed if different to the default
-
-    Note that ``"mpich"`` and ``"openmpi"`` runners have the default command ``"mpirun"``
-    """
-    cores_per_node: Optional[int]
-    """Number of physical CPU cores on a compute node of the platform"""
-
-    logical_cores_per_node: Optional[int]
-    """Number of logical CPU cores on a compute node of the platform"""
-
-    gpus_per_node: Optional[int]
-    """Number of GPU devices on a compute node of the platform"""
-
-    gpu_setting_type: Optional[str]
-    """ How GPUs will be assigned.
-
-    Must take one of the following string options.
-
-    - ``"runner_default"``:       Use default setting for MPI runner (same as if not set).
-    - ``"env"``:                  Use an environment variable (comma-separated list of slots)
-    - ``"option_gpus_per_node"``: Expresses GPUs per node on MPI runner command line.
-    - ``"option_gpus_per_task"``: Expresses GPUs per task on MPI runner command line.
-
-    With the exception of "runner_default", the :attr:`gpu_setting_name`
-    attribute is also required when this attribute is set.
-
-    If "gpu_setting_type" is not provided (same as ``runner_default``) and the
-    MPI runner does not have a default GPU setting in libEnsemble, and no other
-    information is present, then the environment variable ``CUDA_VISIBLE_DEVICES``
-    is used.
-
-    Examples:
-
-    Use environment variable ROCR_VISIBLE_DEVICES to assign GPUs.
-
-    .. code-block:: python
-
-        "gpu_setting_type" = "env"
-        "gpu_setting_name" = "ROCR_VISIBLE_DEVICES"
-
-    Use command line option ``--gpus-per-node``
-
-    .. code-block:: python
-
-        "gpu_setting_type" = "option_gpus_per_node"
-        "gpu_setting_name" = "--gpus-per-node"
-
-    """
-
-    gpu_setting_name: Optional[str]
-    """Name of GPU setting
-
-    See :attr:`gpu_setting_type` for more details.
-
-    """
-
-    gpu_env_fallback: Optional[str]
-    """GPU fallback environment setting if not using an MPI runner.
-
-    For example:
-
-    .. code-block:: python
-
-        "gpu_setting_type" = "runner_default"
-        "gpu_env_fallback" = "ROCR_VISIBLE_DEVICES"
-
-    This example will use the MPI runner default settings when using an MPI runner, but
[…]
