### 2023-11-21 06f4889
universal platforms.py
