### 2023-11-22 3eba099
remove default in/out from tests using default sixhumpcamel
