+++ libensemble/tests/functionality_tests/test_1d_uniform_sampling_with_comm_dup.py
@@ -43,8 +43,6 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
     }
 
     gen_specs = {
