+++ libensemble/tests/functionality_tests/test_elapsed_time_abort.py
@@ -29,8 +29,6 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
         "user": {"pause_time": 2},
     }
 
