+++ libensemble/tests/regression_tests/test_persistent_aposmm_nlopt.py
@@ -47,8 +47,6 @@ if __name__ == "__main__":
     n = 2
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
     }
 
     gen_out = [
