+++ libensemble/tests/functionality_tests/test_1d_subcomm.py
@@ -41,8 +41,6 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
     }
 
     gen_specs = {
