+++ libensemble/tests/functionality_tests/test_1d_splitcomm.py
@@ -37,8 +37,6 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
     }
 
     gen_specs = {
