+++ libensemble/tests/functionality_tests/test_active_persistent_worker_abort.py
@@ -39,8 +39,6 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
     }
 
     gen_out += [("x", float, 2), ("x_on_cube", float, 2)]
