+++ libensemble/tests/functionality_tests/test_sim_input_dir_option.py
@@ -49,7 +49,6 @@ if __name__ == "__main__":
 
     gen_specs = {
         "gen_f": gen_f,
-        "out": [("x", float, (1,))],
         "user": {
             "gen_batch_size": 20,
             "lb": np.array([-3]),
