+++ libensemble/tests/functionality_tests/test_elapsed_time_abort.py
@@ -34,8 +34,6 @@ if __name__ == "__main__":
 
     gen_specs = {
         "gen_f": gen_f,
-        "in": ["sim_id"],
-        "out": [("x", float, (2,))],
         "user": {
             "gen_batch_size": 5,
             "lb": np.array([-3, -2]),
