+++ libensemble/tests/functionality_tests/test_mpi_runners_subnode.py
@@ -82,8 +82,6 @@ if __name__ == "__main__":
 
     gen_specs = {
         "gen_f": gen_f,
-        "in": [],
-        "out": [("x", float, (n,))],
         "user": {
             "gen_batch_size": 20,
             "lb": np.array([-3, -2]),
