+++ libensemble/tests/functionality_tests/test_cancel_in_alloc.py
@@ -41,8 +41,6 @@ if __name__ == "__main__":
 
     gen_specs = {
         "gen_f": gen_f,
-        "in": ["sim_id"],
-        "out": [("x", float, (2,))],
         "user": {
             "gen_batch_size": 5,
             "lb": np.array([-3, -2]),
