+++ libensemble/tests/functionality_tests/test_executor_hworld_timeout.py
@@ -66,8 +66,6 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float), ("cstat", int)],
         "user": {
             "cores": cores_per_task,
             "elapsed_timeout": True,
@@ -76,8 +74,6 @@ if __name__ == "__main__":
 
     gen_specs = {
         "gen_f": gen_f,
-        "in": ["sim_id"],
-        "out": [("x", float, (2,))],
         "user": {
             "lb": np.array([-3, -2]),
             "ub": np.array([3, 2]),
