+++ libensemble/tests/functionality_tests/test_zero_resource_workers_subnode.py
@@ -90,8 +90,6 @@ if __name__ == "__main__":
 
     gen_specs = {
         "gen_f": gen_f,
-        "in": [],
-        "out": [("x", float, (n,))],
         "user": {
             "initial_batch_size": 20,
             "lb": np.array([-3, -2]),
