+++ libensemble/tests/scaling_tests/forces/forces_adv/run_libe_forces.py
@@ -63,7 +63,6 @@ sim_specs = {
 # State the generating function, its arguments, output, and necessary parameters.
 gen_specs = {
     "gen_f": gen_f,  # Generator function
-    "in": [],  # Generator input
     "out": [("x", float, (1,))],  # Name, type and size of data produced (must match sim_specs 'in')
     "user": {
         "lb": np.array([0]),  # Lower bound for random sample array (1D)
