+++ libensemble/tests/scaling_tests/forces/forces_simple/run_libe_forces.py
@@ -41,7 +41,6 @@ if __name__ == "__main__":
 
     ensemble.gen_specs = GenSpecs(
         gen_f=gen_f,
-        inputs=[],  # No input when start persistent generator
         persis_in=["sim_id"],  # Return sim_ids of evaluated points to generator
         outputs=[("x", float, (1,))],
         user={
