+++ libensemble/tests/functionality_tests/test_persistent_uniform_sampling_async.py
@@ -45,8 +45,6 @@ if __name__ == "__main__":
 
     gen_specs = {
         "gen_f": gen_f,
-        "persis_in": ["f", "x", "sim_id"],
-        "out": [("x", float, (n,))],
         "user": {
             "initial_batch_size": nworkers,  # Ensure > 1 alloc to send all sims
             "lb": np.array([-3, -2]),
