+++ libensemble/tests/functionality_tests/test_persistent_sim_uniform_sampling.py
@@ -54,9 +54,6 @@ if __name__ == "__main__":
 
     gen_specs = {
         "gen_f": gen_f,
-        "in": [],
-        "persis_in": ["sim_id", "f", "grad"],
-        "out": [("x", float, (n,))],
         "user": {
             "initial_batch_size": 5,
             "lb": np.array([-3, -2]),
