+++ libensemble/specs.py
@@ -95,14 +95,10 @@ class SimSpecs(BaseModel):
 
     @root_validator
     def set_in_out_from_attrs(cls, values):
-        if not values.get("sim_f"):
-            from libensemble.sim_funcs.one_d_func import one_d_example
-
-            values["sim_f"] = one_d_example
         if hasattr(values.get("sim_f"), "inputs") and not values.get("inputs"):
             values["inputs"] = values.get("sim_f").inputs
         if hasattr(values.get("sim_f"), "outputs") and not values.get("outputs"):
-            values["out"] = values.get("sim_f").outputs
+            values["outputs"] = values.get("sim_f").outputs
         if hasattr(values.get("sim_f"), "persis_in") and not values.get("persis_in"):
             values["persis_in"] = values.get("sim_f").persis_in
         return values
@@ -169,14 +165,10 @@ class GenSpecs(BaseModel):
 
     @root_validator
     def set_in_out_from_attrs(cls, values):
-        if not values.get("gen_f"):
-            from libensemble.gen_funcs.sampling import latin_hypercube_sample
-
-            values["gen_f"] = latin_hypercube_sample
         if hasattr(values.get("gen_f"), "inputs") and not values.get("inputs"):
             values["inputs"] = values.get("gen_f").inputs
         if hasattr(values.get("gen_f"), "outputs") and not values.get("outputs"):
-            values["out"] = values.get("gen_f").outputs
+            values["outputs"] = values.get("gen_f").outputs
         if hasattr(values.get("gen_f"), "persis_in") and not values.get("persis_in"):
             values["persis_in"] = values.get("gen_f").persis_in
         return values
