### 2023-11-22 3ce4d1d
write_sim_func
