+++ libensemble/tests/functionality_tests/test_sim_dirs_with_gen_dirs.py
@@ -58,13 +58,10 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
     }
 
     gen_specs = {
         "gen_f": gen_f,
-        "out": [("x", float, (1,))],
         "user": {
             "gen_batch_size": 20,
             "lb": np.array([-3]),
