+++ libensemble/tests/functionality_tests/test_workflow_dir.py
@@ -44,8 +44,6 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
     }
 
     gen_specs = {
