+++ libensemble/tests/functionality_tests/test_sim_dirs_per_worker.py
@@ -47,8 +47,6 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
     }
 
     gen_specs = {
