+++ libensemble/tests/functionality_tests/test_sim_dirs_with_exception.py
@@ -39,8 +39,6 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
     }
 
     gen_specs = {
