+++ libensemble/tests/functionality_tests/test_sim_dirs_with_gen_dirs.py
@@ -62,6 +62,7 @@ if __name__ == "__main__":
 
     gen_specs = {
         "gen_f": gen_f,
+        "out": [("x", float, (1,))],
         "user": {
             "gen_batch_size": 20,
             "lb": np.array([-3]),
