### 2023-11-24 f168b77
Allow overwriting history file on complete
