+++ libensemble/manager.py
@@ -280,12 +280,11 @@ class Manager:
 
     def _save_every_k(self, fname: str, count: int, k: int, complete: bool) -> None:
         """Saves history every kth step"""
-        if not complete:
-            count = k * (count // k)
+        count = k * (count // k)
         date_start = self._get_date_start_str()
 
         filename = fname.format(self.libE_specs["H_file_prefix"], date_start, count)
-        if not os.path.isfile(filename) and count > 0:
+        if (not os.path.isfile(filename) and count > 0) or complete:
             for old_file in glob.glob(fname.format(self.libE_specs["H_file_prefix"], date_start, "*")):
                 os.remove(old_file)
             np.save(filename, self.hist.trim_H())
