### 2023-11-26 #6031 @SergeyYakubov
@jbytecode - I've completed my review. I think the paper is well-written, and the software itself is promising. It is somewhat similar to what Apache Spark & Co does in the cloud, but more lightweight and developed with HPC in mind (I was not able to verify it works properly on a large supercomputer, though).

Thank you for inviting me.
