### 2023-11-27 88ce933
didn't need to do this kind of "passthrough" since the nworkers property returns the correct value
