### 2023-11-27 24da917
Rename detect_env to detect_comms_env
