### 2023-11-27 8aa1890
with Pydantic2, we need to use merge_field_infos and model_rebuild to retain field defaults and aliases
