### 2023-11-27 c915bb1
using create_model to create new model definitions with validators; setting validator attributes directly didn't work since they didn't get bound to fields. various other cross-version fixes and improvements
