### 2023-11-27 6044725
Allow overwriting history file on completion (#1177)
* Allow overwriting history file on complete

* Restore if

* Add test

* Fix test
