### 2023-11-27 c9b0e86
workflow_dir checking bugfix for pydantic2 validator version
