### 2023-11-27 f107c66
Fix two bugs in Ensemble class (#1162)
* ensure that if we did a temporary cast of a new libE_specs to dict, it gets cast back to a class upon setting. Passthrough nworkers attribute upon setting libE_specs if the Ensemble doesn't have it, but libE_specs does

* fixes and improvements for libE_specs setting and nworkers attributes passthrough, plus a test

* didn't need to do this kind of "passthrough" since the nworkers property returns the correct value
