+++ libensemble/tests/unit_tests/test_ensemble.py
@@ -13,9 +13,29 @@ def test_ensemble_init():
 
     e = Ensemble(parse_args=True)
     assert hasattr(e.libE_specs, "comms"), "internal parse_args() didn't populate defaults for class's libE_specs"
+    assert hasattr(e, "nworkers"), "nworkers should've passed from libE_specs to Ensemble class"
     assert e.is_manager, "parse_args() didn't populate defaults for class's libE_specs"
 
     assert e.logger.get_level() == 20, "Default log level should be 20."
+
+
+def test_ensemble_parse_args_false():
+    from libensemble.ensemble import Ensemble
+    from libensemble.specs import LibeSpecs
+
+    e = Ensemble()  # parse_args defaults to False
+    e.libE_specs = {"comms": "local", "nworkers": 4}
+    assert hasattr(e, "nworkers"), "nworkers should've passed from libE_specs to Ensemble class"
+    assert isinstance(e.libE_specs, LibeSpecs), "libE_specs should've been cast to class"
+
+    # test pass attribute as dict
+    e = Ensemble(libE_specs={"comms": "local", "nworkers": 4})
+    assert hasattr(e, "nworkers"), "nworkers should've passed from libE_specs to Ensemble class"
+    assert isinstance(e.libE_specs, LibeSpecs), "libE_specs should've been cast to class"
+
+    # test that adjusting Ensemble.nworkers also changes libE_specs
+    e.nworkers = 8
+    assert e.libE_specs.nworkers == 8, "libE_specs nworkers not adjusted"
 
 
 def test_from_files():
@@ -91,6 +111,7 @@ def test_full_workflow():
         ),
         exit_criteria=ExitCriteria(gen_max=101),
     )
+
     ens.add_random_streams()
     ens.run()
     if ens.is_manager:
@@ -146,6 +167,7 @@ def test_flakey_workflow():
 
 if __name__ == "__main__":
     test_ensemble_init()
+    test_ensemble_parse_args_false()
     test_from_files()
     test_bad_func_loads()
     test_full_workflow()
