### 2023-11-28 39f6af2
what in tarnation, are we missing a dependency for compilation somehow?
