+++ .github/workflows/basic.yml
@@ -69,7 +69,7 @@ jobs:
             pip install -I --upgrade certifi
 
         - name: Install Ubuntu compilers
-          if: matrix.os == 'ubuntu-latest' && steps.cache.outputs.cache-hit != 'true'
+          if: matrix.os == 'ubuntu-latest'
           run: conda install gcc_linux-64
 
         # Roundabout solution on macos for proper linking with mpicc
@@ -90,7 +90,6 @@ jobs:
             conda install mpi4py ${{ matrix.mpi-version }}
 
         - name: Install generator dependencies
-          if: matrix.os != 'windows-latest' && steps.cache.outputs.cache-hit != 'true'
           run: |
             python -m pip install --upgrade pip
             pip install mpmath
