### 2023-11-28 112253e
fix pydantic2 validator bug with enabling final_save when save_every_k is enabled
