### 2023-11-28 10b1378
adjust required pydantic version, add jobs for testing old-pydantic
