+++ .github/workflows/basic.yml
@@ -18,6 +18,7 @@ jobs:
                 os: [ubuntu-latest]
                 mpi-version: [mpich]
                 python-version: [3.9, "3.10", "3.11", "3.12"]
+                pydantic-version: ["2.5.2"]
                 comms-type: [m, l]
                 include:
                     - os: macos-latest
@@ -27,6 +28,12 @@ jobs:
                     - os: macos-latest
                       python-version: 3.11
                       mpi-version: "mpich=4.0.3"
+                      comms-type: l
+                    - python-version: 3.10
+                      pydantic-version: ["1.10.13"]
+                      comms-type: m
+                    - python-version: 3.10
+                      pydantic-version: ["1.10.13"]
                       comms-type: l
 
         env:
@@ -107,8 +114,9 @@ jobs:
               /usr/share/miniconda3/include
             key: libe-${{ github.ref_name }}-${{ matrix.python-version }}-${{ matrix.comms-type }}
 
-        - name: Install libEnsemble, flake8, lock environment
+        - name: Install libEnsemble, flake8
           run: |
+            pip install pydantic==${{ matrix.pydantic-version }}
             pip install -e .
             flake8 libensemble
 
