+++ .github/workflows/basic.yml
@@ -30,10 +30,10 @@ jobs:
                       mpi-version: "mpich=4.0.3"
                       comms-type: l
                     - python-version: 3.10
-                      pydantic-version: ["1.10.13"]
+                      pydantic-version: "1.10.13"
                       comms-type: m
                     - python-version: 3.10
-                      pydantic-version: ["1.10.13"]
+                      pydantic-version: "1.10.13"
                       comms-type: l
 
         env:
