### 2023-11-28 e2a9bd8
more matrix fixes
