+++ .github/workflows/basic.yml
@@ -22,21 +22,23 @@ jobs:
                 comms-type: [m, l]
                 include:
                     - os: macos-latest
-                      python-version: 3.11
+                      python-version: "3.11"
                       mpi-version: "mpich=4.0.3"
+                      pydantic-version: "2.5.2"
                       comms-type: m
                     - os: macos-latest
-                      python-version: 3.11
+                      python-version: "3.11"
                       mpi-version: "mpich=4.0.3"
+                      pydantic-version: "2.5.2"
                       comms-type: l
                     - os: ubuntu-latest
                       mpi-version: mpich
-                      python-version: 3.10
+                      python-version: "3.10"
                       pydantic-version: "1.10.13"
                       comms-type: m
                     - os: ubuntu-latest
                       mpi-version: mpich
-                      python-version: 3.10
+                      python-version: "3.10"
                       pydantic-version: "1.10.13"
                       comms-type: l
 
