### 2023-11-28 c44ecd1
use cross-version wrapper function instead of model_dump in unit test setup
