+++ libensemble/tests/unit_tests/setup.py
@@ -2,6 +2,7 @@ import numpy as np
 
 from libensemble.history import History
 from libensemble.specs import AllocSpecs
+from libensemble.utils.misc import specs_dump
 
 # -------------------------------------------------------------------------------------------------
 # Set up sim_specs, gen_specs, exit_criteria
@@ -50,7 +51,7 @@ def make_criteria_and_specs_1A(simx=10):
 # Set up history array
 def hist_setup1(sim_max=10, n=1, H0_in=[]):
     sim_specs, gen_specs, exit_criteria = make_criteria_and_specs_0(simx=sim_max, n=n)
-    alloc_specs = AllocSpecs().model_dump()
+    alloc_specs = specs_dump(AllocSpecs())
     H0 = H0_in
     hist = History(alloc_specs, sim_specs, gen_specs, exit_criteria, H0)
     return hist, sim_specs, gen_specs, exit_criteria, alloc_specs
@@ -58,7 +59,7 @@ def hist_setup1(sim_max=10, n=1, H0_in=[]):
 
 def hist_setup2(sim_max=10, H0_in=[]):
     sim_specs, gen_specs, exit_criteria = make_criteria_and_specs_1(simx=sim_max)
-    alloc_specs = AllocSpecs().model_dump()
+    alloc_specs = specs_dump(AllocSpecs())
     H0 = H0_in
     hist = History(alloc_specs, sim_specs, gen_specs, exit_criteria, H0)
     return hist, sim_specs, gen_specs, exit_criteria, alloc_specs
@@ -66,7 +67,7 @@ def hist_setup2(sim_max=10, H0_in=[]):
 
 def hist_setup2A_genout_sim_ids(sim_max=10):
     sim_specs, gen_specs, exit_criteria = make_criteria_and_specs_1A(simx=sim_max)
-    alloc_specs = AllocSpecs().model_dump()
+    alloc_specs = specs_dump(AllocSpecs())
     H0 = []
     hist = History(alloc_specs, sim_specs, gen_specs, exit_criteria, H0)
     return hist, sim_specs, gen_specs, exit_criteria, alloc_specs
