+++ libensemble/tests/unit_tests/test_models.py
@@ -51,7 +51,10 @@ def test_sim_gen_alloc_exit_specs_invalid():
     }
 
     try:
-        SimSpecs.model_validate(bad_specs)
+        if pydanticV1:
+            SimSpecs.parse_obj(bad_specs)
+        elif pydanticV2:
+            SimSpecs.model_validate(bad_specs)
         flag = 0
     except ValidationError as e:
         if pydanticV1:
@@ -62,7 +65,10 @@ def test_sim_gen_alloc_exit_specs_invalid():
     assert flag, "SimSpecs didn't raise ValidationError on invalid specs"
 
     try:
-        GenSpecs.model_validate(bad_specs)
+        if pydanticV1:
+            GenSpecs.model_validate(bad_specs)
+        elif pydanticV2:
+            GenSpecs.parse_obj(bad_specs)
         flag = 0
     except ValidationError as e:
         if pydanticV1:
@@ -75,7 +81,10 @@ def test_sim_gen_alloc_exit_specs_invalid():
     bad_ec = {"stop_vals": 0.5}
 
     try:
-        ExitCriteria.model_validate(bad_ec)
+        if pydanticV1:
+            ExitCriteria.parse_obj(bad_ec)
+        elif pydanticV2:
+            ExitCriteria.model_validate(bad_ec)
         flag = 0
     except ValidationError:
         flag = 1
@@ -85,22 +94,41 @@ def test_sim_gen_alloc_exit_specs_invalid():
 def test_libe_specs():
     sim_specs, gen_specs, exit_criteria = setup.make_criteria_and_specs_0()
     libE_specs = {"mpi_comm": Fake_MPI(), "comms": "mpi"}
-    ls = LibeSpecs.model_validate(libE_specs)
+    if pydanticV1:
+        ls = LibeSpecs.parse_obj(libE_specs)
+    elif pydanticV2:
+        ls = LibeSpecs.model_validate(libE_specs)
 
     libE_specs["sim_input_dir"] = "./simdir"
     libE_specs["sim_dir_copy_files"] = ["./simdir"]
-    ls = LibeSpecs.model_validate(libE_specs)
+    if pydanticV1:
+        ls = LibeSpecs.parse_obj(libE_specs)
+    elif pydanticV2:
+        ls = LibeSpecs.model_validate(libE_specs)
 
-    ls = LibeSpecs.model_validate({"comms": "tcp", "nworkers": 4})
+    libE_specs = {"comms": "tcp", "nworkers": 4}
+
+    if pydanticV1:
+        ls = LibeSpecs.parse_obj(libE_specs)
+    elif pydanticV2:
+        ls = LibeSpecs.model_validate(libE_specs)
     assert ls.disable_resource_manager, "resource manager should be disabled when using tcp comms"
-    ls = LibeSpecs.model_validate({"comms": "tcp", "workers": ["hello.host"]})
+
+    libE_specs = {"comms": "tcp", "workers": ["hello.host"]}
+    if pydanticV1:
+        ls = LibeSpecs.parse_obj(libE_specs)
+    elif pydanticV2:
+        ls = LibeSpecs.model_validate(libE_specs)
 
 
 def test_libe_specs_invalid():
     bad_specs = {"comms": "local", "zero_resource_workers": 2, "sim_input_dirs": ["obj"]}
 
     try:
-        LibeSpecs.model_validate(bad_specs)
+        if pydanticV1:
+            LibeSpecs.parse_obj(bad_specs)
+        elif pydanticV2:
+            LibeSpecs.model_validate(bad_specs)
         flag = 0
     except ValidationError as e:
         assert len(e.errors()) == 2, "LibeSpecs model should have detected 2 errors in specs"
