+++ libensemble/tests/unit_tests/test_models.py
@@ -58,7 +58,7 @@ def test_sim_gen_alloc_exit_specs_invalid():
         flag = 0
     except ValidationError as e:
         if pydanticV1:
-            assert len(e.errors()) == 5, "SimSpecs model should have detected 5 errors in specs"
+            assert len(e.errors()) == 4, "SimSpecs model should have detected 4 errors in specs"
         elif pydanticV2:  # Pydantic 2 will point out individual errors in iterables
             assert len(e.errors()) == 7, "SimSpecs model should have detected 7 errors in specs"
         flag = 1
@@ -131,7 +131,7 @@ def test_libe_specs_invalid():
             LibeSpecs.model_validate(bad_specs)
         flag = 0
     except ValidationError as e:
-        assert len(e.errors()) == 2, "LibeSpecs model should have detected 2 errors in specs"
+        assert len(e.errors()) == 1, "LibeSpecs model should have detected errors in specs"
         flag = 1
     assert flag, "LibeSpecs didn't raise ValidationError on invalid specs"
 
