+++ libensemble/tests/unit_tests/test_models.py
@@ -131,7 +131,7 @@ def test_libe_specs_invalid():
             LibeSpecs.model_validate(bad_specs)
         flag = 0
     except ValidationError as e:
-        assert len(e.errors()) == 1, "LibeSpecs model should have detected errors in specs"
+        assert len(e.errors()) == 2, "LibeSpecs model should have detected errors in specs"
         flag = 1
     assert flag, "LibeSpecs didn't raise ValidationError on invalid specs"
 
