### 2023-11-28 6c3345e
really dont know why the number of errors is variable, but at least its erroring and displaying multiple?
