### 2023-11-29 468845f
change these unsets back to del, adjust cache
