+++ libensemble/utils/validators.py
@@ -1,6 +1,6 @@
 import secrets
 from pathlib import Path
-from typing import Callable, Union
+from typing import Callable
 
 import numpy as np
 
@@ -18,54 +18,81 @@ _UFUNC_INVALID_ERR = "Specified sim_f or gen_f is not callable. It should be a u
 _OUT_DTYPE_ERR = "unable to coerce into a NumPy dtype. It should be a list of 2-tuples or 3-tuples"
 _IN_INVALID_ERR = "value should be a list of field names (a list of strings)"
 
-""" Pydantic validation logic, implemented in both V1 and V2"""
+
+def check_valid_out(cls, v):
+    try:
+        _ = np.dtype(v)
+    except TypeError:
+        raise ValueError(_OUT_DTYPE_ERR.format(v))
+    else:
+        return v
+
+
+def check_valid_in(cls, v):
+    if not all(isinstance(s, str) for s in v):
+        raise ValueError(_IN_INVALID_ERR)
+    return v
+
+
+def check_valid_comms_type(cls, value):
+    assert value in ["mpi", "local", "local_threading", "tcp"], "Invalid comms type"
+    return value
+
+
+def set_platform_specs_to_class(cls, value) -> Platform:
+    if isinstance(value, dict):
+        value = Platform(**value)
+    return value
+
+
+def check_input_dir_exists(cls, value):
+    if value:
+        if isinstance(value, str):
+            value = Path(value).absolute()
+        assert value.exists(), "value does not refer to an existing path"
+        assert value != Path("."), "Value can't refer to the current directory ('.' or Path('.'))."
+    return value
+
+
+def check_inputs_exist(cls, value):
+    value = [Path(path).absolute() for path in value]
+    for f in value:
+        assert f.exists(), f"'{f}' in Value does not refer to an existing path."
+    return value
+
+
+def check_gpu_setting_type(cls, value):
+    if value is not None:
+        assert value in [
+            "runner_default",
+            "env",
+            "option_gpus_per_node",
+            "option_gpus_per_task",
+        ], "Invalid label for GPU specification type"
+    return value
+
+
+def check_mpi_runner_type(cls, value):
+    if value is not None:
+        assert value in ["mpich", "openmpi", "aprun", "srun", "jsrun", "msmpi", "custom"], "Invalid MPI runner name"
+    return value
+
 
 if pydanticV1:
     from pydantic import root_validator, validator
 
     # SPECS VALIDATORS #####
 
-    @validator("outputs", pre=True)
-    def check_valid_out(cls, v):
-        try:
-            _ = np.dtype(v)
-        except TypeError:
-            raise ValueError(_OUT_DTYPE_ERR.format(v))
-        else:
-            return v
-
-    @validator("inputs", "persis_in", pre=True)
-    def check_valid_in(cls, v):
-        if not all(isinstance(s, str) for s in v):
-            raise ValueError(_IN_INVALID_ERR)
-        return v
-
-    @validator("comms")
-    def check_valid_comms_type(cls, value):
-        assert value in ["mpi", "local", "local_threading", "tcp"], "Invalid comms type"
-        return value
-
-    @validator("platform_specs")
-    def set_platform_specs_to_class(cls, value) -> Platform:
-        if isinstance(value, dict):
-            value = Platform(**value)
-        return value
-
-    @validator("sim_input_dir", "gen_input_dir")
-    def check_input_dir_exists(cls, value):
-        if value:
-            if isinstance(value, str):
-                value = Path(value).absolute()
-            assert value.exists(), "value does not refer to an existing path"
-            assert value != Path("."), "Value can't refer to the current directory ('.' or Path('.'))."
-        return value
-
[…]
