### 2023-11-29 1a041e7
Dealing with the weird APOSMM r_k case when there is only one sample
