### 2023-11-29 54ac7cc
moving some logic to specs_checkers
