### 2023-11-29 ecf556d
removing cache-hit condition
