+++ libensemble/tests/regression_tests/test_persistent_aposmm_ibcdfo_pounders.py
@@ -1,7 +1,17 @@
 """
-Runs libEnsemble with APOSMM+IBCDFO on the chwirut least-squares problem.
-All 214 residual calculations for a given point are performed as a single
-simulation evaluation.
+Runs libEnsemble with APOSMM+IBCDFO on two test problems. Only a single
+optimization run is being performed for the below setup.
+
+The first case uses POUNDERS to solve the chwirut least-squares problem. For
+this case, all chwirut 214 residual calculations for a given point are
+performed as a single simulation evaluation.
+
+The second case uses the generalized POUNDERS to minimize normalized beamline
+emittance. The "beamline simulation" is a synthetic polynomial test function
+that takes in 4 variables and returning 3 outputs. These outputs represent
+position <x>, momentum <p_x>, and the correlation between them <x p_x>.
+
+These values are then mapped to the normalized emittance <x> <p_x> - <x p_x>.
 
 Execute via one of the following commands:
    mpiexec -np 3 python test_persistent_aposmm_ibcdfo.py
@@ -21,7 +31,7 @@ import numpy as np
 
 import libensemble.gen_funcs
 from libensemble.libE import libE
-from libensemble.sim_funcs.chwirut1 import chwirut_eval as sim_f
+from libensemble.sim_funcs.chwirut1 import chwirut_eval
 
 libensemble.gen_funcs.rc.aposmm_optimizers = "ibcdfo"
 
@@ -31,6 +41,7 @@ from libensemble.tools import add_unique_random_streams, parse_args, save_libE_o
 
 try:
     from ibcdfo.pounders import pounders  # noqa: F401
+    from ibcdfo.pounders.general_h_funs import emittance_combine, emittance_h
 except ModuleNotFoundError:
     sys.exit("Please 'pip install ibcdfo'")
 
@@ -41,8 +52,22 @@ except ModuleNotFoundError:
     sys.exit("Ensure https://github.com/POptUS/minq has been cloned and that minq/py/minq5/ is on the PYTHONPATH")
 
 
-def combine_component(x):
+def sum_squared(x):
     return np.sum(np.power(x, 2))
+
+
+def synthetic_beamline_mapping(H, _, sim_specs):
+    x = H["x"][0]
+    assert len(x) == 4, "Assuming 4 inputs to this function"
+    y = np.zeros(3)  # Synthetic beamline outputs
+    y[0] = x[0] ** 2 + 1.0
+    y[1] = x[1] ** 2 + 2.0
+    y[2] = x[2] * x[3] + 0.5
+
+    Out = np.zeros(1, dtype=sim_specs["out"])
+    Out["fvec"] = y
+    Out["f"] = y[0] * y[1] - y[2] ** 2
+    return Out
 
 
 # Main block is necessary only when using local comms with spawn start method (default on macOS and Windows).
@@ -53,74 +78,62 @@ if __name__ == "__main__":
 
     assert nworkers == 2, "This test is just for two workers"
 
-    # Declare the run parameters/functions
-    m = 214
-    n = 3
+    for inst in range(2):
+        if inst == 0:
+            # Declare the run parameters/functions
+            m = 214
+            n = 3
+            sim_f = chwirut_eval
+        elif inst == 1:
+            m = 3
+            n = 4
+            sim_f = synthetic_beamline_mapping
 
-    sim_specs = {
-        "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float), ("fvec", float, m)],
-        "user": {
-            "combine_component_func": combine_component,
-        },
-    }
+        sim_specs = {
+            "sim_f": sim_f,
+            "in": ["x"],
+            "out": [("f", float), ("fvec", float, m)],
+        }
 
-    gen_out = [
[…]
