### 2023-11-30 9b75033
Removing unused 'old_runs' persis_info entry
