### 2023-11-30 76d0e92
specs_checker_getattr can now return a default value
