### 2023-12-01 4d2f9ed
Supporting more ways to stop APOSMM
