+++ libensemble/gen_funcs/persistent_aposmm.py
@@ -87,6 +87,12 @@ def aposmm(H, persis_info, gen_specs, libE_info):
       points must satisfy
     - ``'rk_const' [float]``: Multiplier in front of the r_k value
     - ``'max_active_runs' [int]``: Bound on number of runs APOSMM is advancing
+    - ``'stop_after_k_minima' [int]``: Tell APOSMM to stop after this many
+      local minima have been identified by a local optimization run.
+    - ``'stop_after_k_runs' [int]``: Tell APOSMM to stop after this many runs
+      have ended. (The number of ended runs may be less than the number of
+      minima if, for example, a local optimization run ends due to a evaluation
+      constraint, but not convergence criteria.)
 
     If the rules in ``decide_where_to_start_localopt`` produces more than
     ``'max_active_runs'`` in some iteration, then existing runs are prioritized.
@@ -130,7 +136,7 @@ def aposmm(H, persis_info, gen_specs, libE_info):
                       unless opt_flag is 1)
     opt_flag:         1 if the run ended with an optimal point (x_opt) or
                       0 if it ended because e.g., maxiters/maxevals were reached
-    num_samples_needed:   Number of additional uniformly drawn samples needed
+    num_samples:      Number of additional uniformly drawn samples needed
 
 
     Description of persistent variables used to maintain the state of APOSMM
@@ -144,9 +150,16 @@ def aposmm(H, persis_info, gen_specs, libE_info):
         user_specs = gen_specs["user"]
         ps = PersistentSupport(libE_info, EVAL_GEN_TAG)
         n, n_s, rk_const, ld, mu, nu, comm, local_H = initialize_APOSMM(H, user_specs, libE_info)
-        local_opters, sim_id_to_child_inds, run_order, run_pts, total_runs, fields_to_pass = initialize_children(
-            user_specs
-        )
+        (
+            local_opters,
+            sim_id_to_child_inds,
+            run_order,
+            run_pts,
+            total_runs,
+            ended_runs,
+            fields_to_pass,
+        ) = initialize_children(user_specs)
+
         if user_specs["initial_sample_size"] != 0:
             # Send our initial sample. We don't need to check that n_s is large enough:
             # the alloc_func only returns when the initial sample has function values.
@@ -176,8 +189,12 @@ def aposmm(H, persis_info, gen_specs, libE_info):
                     persis_info["run_order"] = run_order
                     break
 
-                if np.sum(local_H["local_min"]) >= user_specs.get("stop_after_this_many_minima", np.inf):
+                if np.sum(local_H["local_min"]) >= user_specs.get("stop_after_k_minima", np.inf) or len(
+                    ended_runs
+                ) >= user_specs.get("stop_after_k_runs", np.inf):
                     # This break happens here so the manager can be informed about the last minima.
+                    clean_up_and_stop(local_opters)
+                    persis_info["run_order"] = run_order
                     break
 
                 n_s, n_r = update_local_H_after_receiving(local_H, n, n_s, user_specs, Work, calc_in, fields_to_pass)
@@ -193,6 +210,7 @@ def aposmm(H, persis_info, gen_specs, libE_info):
                                 opt_ind = update_history_optimal(x_opt, opt_flag, local_H, run_order[child_idx])
                                 new_opt_inds_to_send_mgr.append(opt_ind)
                                 local_opters.pop(child_idx)
+                                ended_runs.append(child_idx)
                             else:
                                 add_to_local_H(local_H, x_new, user_specs, local_flag=1, on_cube=True)
[…]
