### 2023-12-01 255efc9
Passing run_max_eval to IBCDFO APOSMM runs
