### 2023-12-01 7d0718e
add a History test where an entry dtype is "object"
