+++ .github/workflows/basic.yml
@@ -90,7 +90,6 @@ jobs:
             conda install mpi4py ${{ matrix.mpi-version }}
 
         - name: Install generator dependencies
-          if: matrix.os != 'windows-latest' && steps.cache.outputs.cache-hit != 'true'
           run: |
             python -m pip install --upgrade pip
             pip install mpmath
