+++ .github/workflows/extra.yml
@@ -95,7 +95,6 @@ jobs:
             pip install mpi4py
 
         - name: Install generator dependencies
-          if: steps.cache.outputs.cache-hit != 'true'
           run: |
             python -m pip install --upgrade pip
             conda env update --file install/gen_deps_environment.yml
