### 2023-12-01 04da3b1
Removing another cache check
