### 2023-12-01 106dead
decorate latin_hypercube_sample
