+++ libensemble/tests/functionality_tests/test_fast_alloc.py
@@ -34,7 +34,6 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
         "out": [("f", float), ("large", float, 1000000)],
         "user": {},
     }
