+++ libensemble/tests/functionality_tests/test_persistent_uniform_sampling_async.py
@@ -38,8 +38,6 @@ if __name__ == "__main__":
     n = 2
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
         "user": {"uniform_random_pause_ub": 0.5},
     }
 
