+++ libensemble/tests/functionality_tests/test_1d_super_simple.py
@@ -19,15 +19,19 @@ from libensemble.gen_funcs.sampling import latin_hypercube_sample as gen_f
 
 # Import libEnsemble items for this test
 from libensemble.libE import libE
+from libensemble.specs import input_fields, output_data
 from libensemble.tools import add_unique_random_streams, parse_args, save_libE_output
 
 
+@input_fields(["x"])
+@output_data([("f", float)])
 def sim_f(In):
     Out = np.zeros(1, dtype=[("f", float)])
     Out["f"] = np.linalg.norm(In)
     return Out
 
 
+@input_fields(["x"])
 def sim_f_noreturn(In):
     print(np.linalg.norm(In))
 
@@ -37,8 +41,6 @@ if __name__ == "__main__":
 
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
     }
 
     gen_specs = {
@@ -64,7 +66,6 @@ if __name__ == "__main__":
     # Test running a sim_f without any returns
     sim_specs = {
         "sim_f": sim_f_noreturn,
-        "in": ["x"],
     }
 
     H, persis_info, flag = libE(sim_specs, gen_specs, exit_criteria, persis_info, libE_specs=libE_specs)
