+++ libensemble/tests/functionality_tests/test_persistent_uniform_gen_decides_stop.py
@@ -41,8 +41,6 @@ if __name__ == "__main__":
 
         sim_specs = {
             "sim_f": sim_f,
-            "in": ["x"],
-            "out": [("f", float)],
             "user": {"uniform_random_pause_ub": 0.5},
         }
 
