### 2023-12-01 7ced813
decorate gpu_variable_resources_from_gen
