+++ libensemble/tests/functionality_tests/test_mpi_gpu_settings_mock_nodes_multi_task.py
@@ -58,8 +58,6 @@ if __name__ == "__main__":
     n = 2
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
         "user": {"dry_run": True},
     }
 
