+++ libensemble/tests/regression_tests/test_GPU_variable_resources.py
@@ -61,8 +61,6 @@ if __name__ == "__main__":
 
     gpu_test.sim_specs = SimSpecs(
         sim_f=sim_f,
-        inputs=["x"],
-        out=[("f", float)],
         user={"dry_run": False},
     )
     gpu_test.gen_specs = GenSpecs(
