+++ libensemble/tests/functionality_tests/test_GPU_gen_resources.py
@@ -70,8 +70,6 @@ if __name__ == "__main__":
     n = 2
     sim_specs = {
         "sim_f": sim_f,
-        "in": ["x"],
-        "out": [("f", float)],
         "user": {"dry_run": dry_run},
     }
 
