### 2023-12-01 ecf32e6
Moving local sine tutorial content
