### 2023-12-01 ad1a185
Add Aurora to known platforms
