### 2023-12-04 37dd599
Docs/testing tutorial content (#1182)
* Moving local sine tutorial content

* Adding tutorial content

* Moving to functionality tests for now

* adjust suggested filenames to match captions

---------

Co-authored-by: jlnav <jnavarro@anl.gov>
