+++ docs/tutorials/local_sine_tutorial.rst
@@ -61,35 +61,13 @@ need to write a new allocation function.
 
         Later on, we'll populate :class:`gen_specs<libensemble.specs.GenSpecs>` and ``persis_info`` when we initialize libEnsemble.
 
-        For now, create a new Python file named ``gen.py``. Write the following:
+        For now, create a new Python file named ``sine_gen.py``. Write the following:
 
-        .. code-block:: python
+        .. literalinclude:: ../../libensemble/tests/functionality_tests/sine_gen.py
+            :language: python
             :linenos:
-            :caption: examples/tutorials/simple_sine/gen.py
+            :caption: examples/tutorials/simple_sine/sine_gen.py
 
-            import numpy as np
-
-
-            def gen_random_sample(InputArray, persis_info, gen_specs):
-                # Pull out user parameters
-                user_specs = gen_specs["user"]
-
-                # Get lower and upper bounds
-                lower = user_specs["lower"]
-                upper = user_specs["upper"]
-
-                # Determine how many values to generate
-                num = len(lower)
-                batch_size = user_specs["gen_batch_size"]
-
-                # Create empty array of "batch_size" zeros. Array dtype should match "out" fields
-                OutputArray = np.zeros(batch_size, dtype=gen_specs["out"])
-
-                # Set the "x" output field to contain random numbers, using random stream
-                OutputArray["x"] = persis_info["rand_stream"].uniform(lower, upper, (batch_size, num))
-
-                # Send back our output and persis_info
-                return OutputArray, persis_info
 
         Our function creates ``batch_size`` random numbers uniformly distributed
         between the ``lower`` and ``upper`` bounds. A random stream
@@ -103,24 +81,13 @@ need to write a new allocation function.
         The only new parameter here is :ref:`sim_specs<datastruct-sim-specs>`, which
         serves a purpose similar to the :class:`gen_specs<libensemble.specs.GenSpecs>` dictionary.
 
-        Create a new Python file named ``sim.py``. Write the following:
+        Create a new Python file named ``sine_sim.py``. Write the following:
 
-        .. code-block:: python
+        .. literalinclude:: ../../libensemble/tests/functionality_tests/sine_sim.py
+            :language: python
             :linenos:
-            :caption: examples/tutorials/simple_sine/sim.py
+            :caption: examples/tutorials/simple_sine/sine_sim.py
 
-            import numpy as np
-
-
-            def sim_find_sine(InputArray, _, sim_specs):
-                # Create an output array of a single zero
-                OutputArray = np.zeros(1, dtype=sim_specs["out"])
-
-                # Set the zero to the sine of the InputArray value
-                OutputArray["y"] = np.sin(InputArray["x"])
-
-                # Send back our output
-                return OutputArray
 
         Our simulator function is called by a worker for every work item produced by
         the generator function. This function calculates the sine of the passed value,
@@ -139,17 +106,11 @@ need to write a new allocation function.
         specify the number of workers and the manager/worker intercommunication method.
         ``"local"``, refers to Python's multiprocessing.
 
-        .. code-block:: python
+        .. literalinclude:: ../../libensemble/tests/functionality_tests/test_local_sine_tutorial.py
+            :language: python
             :linenos:
+            :end-at: libE_specs = LibeSpecs
 
-            import numpy as np
-            from libensemble import Ensemble
-            from libensemble.specs import LibeSpecs, SimSpecs, GenSpecs, ExitCriteria
-            from gen import gen_random_sample
-            from sim import sim_find_sine
-
-            if __name__ == "__main__":  # Python-quirk required on macOS and windows
-                libE_specs = LibeSpecs(nworkers=4, comms="local")
 
[…]
