### 2023-12-04 Libensemble/libensemble#1167 @jlnav
Closed by #1171
