### 2023-12-04 3cbd84f
Testing / misc fixes (#1184)
* differentiate caches in basic vs. extra workflows, though cache-restoring is disabled for now. run-tests.sh runs a test with comms=local, nworkers=4 if TESTSUITE_COMMS and TESTSUITE_NPROCS are empty

* install matplotlib

* fix import
