### 2023-12-04 bfcd48d
adjust new unit test for pydantic 2
