+++ libensemble/tests/functionality_tests/forces_simf.py
@@ -1,0 +1,42 @@
+import numpy as np
+
+# Optional status codes to display in libE_stats.txt for each gen or sim
+from libensemble.message_numbers import TASK_FAILED, WORKER_DONE
+
+
+def run_forces(H, persis_info, sim_specs, libE_info):
+    """Runs the forces MPI application"""
+
+    calc_status = 0
+
+    # Parse out num particles, from generator function
+    particles = str(int(H["x"][0][0]))
+
+    # app arguments: num particles, timesteps, also using num particles as seed
+    args = particles + " " + str(10) + " " + particles
+
+    # Retrieve our MPI Executor
+    exctr = libE_info["executor"]
+
+    # Submit our forces app for execution.
+    task = exctr.submit(app_name="forces", app_args=args)
+
+    # Block until the task finishes
+    task.wait()
+
+    # Try loading final energy reading, set the sim's status
+    statfile = "forces.stat"
+    try:
+        data = np.loadtxt(statfile)
+        final_energy = data[-1]
+        calc_status = WORKER_DONE
+    except Exception:
+        final_energy = np.nan
+        calc_status = TASK_FAILED
+
+    # Define our output array, populate with energy reading
+    output = np.zeros(1, dtype=sim_specs["out"])
+    output["energy"] = final_energy
+
+    # Return final information to worker, for reporting to manager
+    return output, persis_info, calc_status
