### 2023-12-04 80b0139
first attempt at building forces upon running MPI reg tests
