### 2023-12-05 12da602
Cleanup and moving building of forces in run-tests.sh
