+++ .github/workflows/extra.yml
@@ -108,7 +108,7 @@ jobs:
         - name: Install surmise
           if: matrix.os != 'macos-latest'
           run: |
-            pip install --upgrade git+https://github.com/bandframework/surmise.git@develop
+            pip install --upgrade git+https://github.com/bandframework/surmise.git
 
         - name: Build ytopt and dependencies
           if: matrix.python-version <= '3.10' && matrix.os != 'macos-latest'
