### 2023-12-06 86c8440
implement local threading into parse_argso
