### 2023-12-06 3ef99a8
Trying gpCAM interface
