+++ libensemble/tests/regression_tests/test_gpCAM.py
@@ -1,0 +1,71 @@
+"""
+Tests libEnsemble with gpCAM
+
+Execute via one of the following commands (e.g. 3 workers):
+   mpiexec -np 4 python test_gpCAM.py
+   python test_gpCAM.py --nworkers 3 --comms local
+
+When running with the above commands, the number of concurrent evaluations of
+the objective function will be 2, as one of the three workers will be the
+persistent generator.
+"""
+
+# Do not change these lines - they are parsed by run-tests.sh
+# TESTSUITE_COMMS: mpi local
+# TESTSUITE_NPROCS: 3 4
+
+import sys
+
+import numpy as np
+
+from libensemble.alloc_funcs.start_only_persistent import only_persistent_gens as alloc_f
+from libensemble.gen_funcs.persistent_gpCAM import persistent_gpCAM as gen_f
+
+# Import libEnsemble items for this test
+from libensemble.libE import libE
+from libensemble.sim_funcs.rosenbrock import rosenbrock_eval as sim_f
+from libensemble.tools import add_unique_random_streams, parse_args, save_libE_output
+
+# Main block is necessary only when using local comms with spawn start method (default on macOS and Windows).
+if __name__ == "__main__":
+    nworkers, is_manager, libE_specs, _ = parse_args()
+
+    if nworkers < 2:
+        sys.exit("Cannot run with a persistent worker if only one worker -- aborting...")
+
+    n = 4
+    batch_size = 15
+    num_batches = 10
+
+    sim_specs = {
+        "sim_f": sim_f,
+        "in": ["x"],
+        "out": [
+            ("f", float),
+        ],
+    }
+
+    gen_specs = {
+        "gen_f": gen_f,
+        "persis_in": ["x", "f", "sim_id"],
+        "out": [("x", float, (n,))],
+        "user": {
+            "batch_size": batch_size,
+            "lb": np.array([-3, -2, -1, -1]),
+            "ub": np.array([3, 2, 1, 1]),
+        },
+    }
+
+    alloc_specs = {"alloc_f": alloc_f}
+
+    exit_criteria = {"sim_max": num_batches * batch_size, "wallclock_max": 300}
+
+    persis_info = add_unique_random_streams({}, nworkers + 1)
+
+    # Perform the run
+    H, persis_info, flag = libE(sim_specs, gen_specs, exit_criteria, persis_info, alloc_specs, libE_specs)
+
+    if is_manager:
+        assert len(np.unique(H["gen_ended_time"])) == num_batches
+
+        save_libE_output(H, persis_info, __file__, nworkers)
