### 2023-12-06 0797e01
Adding gpcam to testing environment
