+++ .github/workflows/basic.yml
@@ -101,7 +101,7 @@ jobs:
         - name: Install generator dependencies
           run: |
             python -m pip install --upgrade pip
-            pip install mpmath matplotlib
+            pip install mpmath matplotlib gpcam
             conda install numpy nlopt scipy
 
         - uses: actions/cache/save@v3
