### 2023-12-06 9d619f1
Adding gpcam to testing environment
