+++ .github/workflows/basic.yml
@@ -101,8 +101,12 @@ jobs:
         - name: Install generator dependencies
           run: |
             python -m pip install --upgrade pip
-            pip install mpmath matplotlib gpcam
+            pip install mpmath matplotlib
             conda install numpy nlopt scipy
+
+        - name: Install gpcam
+          run: |
+            pip install gpcam
 
         - uses: actions/cache/save@v3
           name: Save dependencies to cache
