### 2023-12-06 38e715d
adjust dependency listings for autodoc_pydantic and pydantic 1.10 lower bound
