### 2023-12-07 3e03f8d
Refactor/ support both Pydantic 1 and 2 (#1135)
* kinda large initial commit

* avoid recursive-validation issue by validators setting attributes by dict. lock to >= current pydantic

* fix no-attribute by .__dict__.get()

* ConfigDict for Platform model, plus apparently new pydantic warns about issues with "del"ing attributes

* skip more steps in deps cache-hit on CI

* tentative change to support both pydantic 1 and pydantic 2

* attempts to set specs and platforms modules based on pydantic version

* Created submodules that import/set versions of specs, platforms based on detected pydantic
[…]
