+++ libensemble/ensemble.py
@@ -14,6 +14,7 @@ from libensemble.specs import AllocSpecs, ExitCriteria, GenSpecs, LibeSpecs, Sim
 from libensemble.tools import add_unique_random_streams
 from libensemble.tools import parse_args as parse_args_f
 from libensemble.tools import save_libE_output
+from libensemble.utils.misc import specs_dump
 
 ATTR_ERR_MSG = 'Unable to load "{}". Is the function or submodule correctly named?'
 ATTR_ERR_MSG = "\n" + 10 * "*" + ATTR_ERR_MSG + 10 * "*" + "\n"
@@ -325,8 +326,8 @@ class Ensemble:
             return
 
         # Cast new libE_specs temporarily to dict
-        if isinstance(new_specs, LibeSpecs):
-            new_specs = new_specs.dict(by_alias=True, exclude_none=True, exclude_unset=True)
+        if not isinstance(new_specs, dict):
+            new_specs = specs_dump(new_specs, by_alias=True, exclude_none=True, exclude_unset=True)
 
         # Unset "comms" if we already have a libE_specs that contains that field, that came from parse_args
         if new_specs.get("comms") and hasattr(self._libE_specs, "comms") and self.parsed:
@@ -464,14 +465,7 @@ class Ensemble:
 
         if len(userf_fields):
             for f in userf_fields:
-                if f == "inputs":
-                    loaded_spec["in"] = field_f[f](loaded_spec[f])
-                    loaded_spec.pop("inputs")
-                elif f == "outputs":
-                    loaded_spec["out"] = field_f[f](loaded_spec[f])
-                    loaded_spec.pop("outputs")
-                else:
-                    loaded_spec[f] = field_f[f](loaded_spec[f])
+                loaded_spec[f] = field_f[f](loaded_spec[f])
 
         return loaded_spec
 
@@ -487,13 +481,9 @@ class Ensemble:
                     old_spec.pop("inputs")  # avoid clashes
                 elif old_spec.get("out") and old_spec.get("outputs"):
                     old_spec.pop("inputs")  # avoid clashes
-            elif isinstance(old_spec, ClassType):
-                old_spec.__dict__.update(**loaded_spec)
-                old_spec = old_spec.dict(by_alias=True)
+                setattr(self, f, ClassType(**old_spec))
             else:  # None. attribute not set yet
                 setattr(self, f, ClassType(**loaded_spec))
-                return
-            setattr(self, f, ClassType(**old_spec))
 
     def from_yaml(self, file_path: str):
         """Parameterizes libEnsemble from ``yaml`` file"""
