+++ .github/workflows/basic.yml
@@ -18,15 +18,28 @@ jobs:
                 os: [ubuntu-latest]
                 mpi-version: [mpich]
                 python-version: [3.9, "3.10", "3.11", "3.12"]
+                pydantic-version: ["2.5.2"]
                 comms-type: [m, l]
                 include:
                     - os: macos-latest
-                      python-version: 3.11
+                      python-version: "3.11"
                       mpi-version: "mpich=4.0.3"
+                      pydantic-version: "2.5.2"
                       comms-type: m
                     - os: macos-latest
-                      python-version: 3.11
+                      python-version: "3.11"
                       mpi-version: "mpich=4.0.3"
+                      pydantic-version: "2.5.2"
+                      comms-type: l
+                    - os: ubuntu-latest
+                      mpi-version: mpich
+                      python-version: "3.10"
+                      pydantic-version: "1.10.13"
+                      comms-type: m
+                    - os: ubuntu-latest
+                      mpi-version: mpich
+                      python-version: "3.10"
+                      pydantic-version: "1.10.13"
                       comms-type: l
 
         env:
@@ -61,7 +74,7 @@ jobs:
               /usr/share/miniconda3/bin
               /usr/share/miniconda3/lib
               /usr/share/miniconda3/include
-            key: libe-${{ github.ref_name }}-${{ matrix.python-version }}-${{ matrix.comms-type }}-basic
+            key: libe-${{ github.ref_name }}-${{ matrix.python-version }}-${{ matrix.comms-type }}-${{ matrix.pydantic-version }}-basic
 
         - name: Force-update certifi
           run: |
@@ -116,8 +129,9 @@ jobs:
               /usr/share/miniconda3/include
             key: libe-${{ github.ref_name }}-${{ matrix.python-version }}-${{ matrix.comms-type }}
 
-        - name: Install libEnsemble, flake8, lock environment
+        - name: Install libEnsemble, flake8
           run: |
+            pip install pydantic==${{ matrix.pydantic-version }}
             pip install -e .
             flake8 libensemble
 
