### 2023-12-07 d197692
Adding a second gpCAM gen
