### 2023-12-07 832adf8
Moving gpCAM test to extra
