+++ .github/workflows/extra.yml
@@ -97,7 +97,7 @@ jobs:
         - name: Install generator dependencies
           run: |
             python -m pip install --upgrade pip
-            pip install mpmath matplotlib
+            pip install mpmath matplotlib gpcam
             conda env update --file install/gen_deps_environment.yml
 
         - name: Install ax-platform
