### 2023-12-08 1be146e
Removing gpCAM from 3.12
