### 2023-12-08 9911d79
Removing erroring functionality test and extra-test build of gpCAM
