+++ .github/workflows/extra.yml
@@ -97,13 +97,14 @@ jobs:
         - name: Install generator dependencies
           run: |
             python -m pip install --upgrade pip
-            pip install mpmath matplotlib gpcam
+            pip install mpmath matplotlib
             conda env update --file install/gen_deps_environment.yml
 
         - name: Install ax-platform
           if: matrix.python-version != '3.12'
           run: |
             pip install ax-platform==0.2.8
+            pip install gpcam
 
         - name: Install surmise
           if: matrix.os != 'macos-latest'
