### 2023-12-08 17dadce
Removing erroring functionality test
