### 2023-12-08 1cf96c6
ensure we do an exception test with abort_on_exception being true
