### 2023-12-08 733068c
Fixing gpCAM gen
