### 2023-12-08 2d28b7d
removing unneeded commands
