+++ .github/workflows/extra.yml
@@ -151,13 +151,6 @@ jobs:
             # pip install git+https://github.com/dragonfly/dragonfly.git
             pip install git+https://github.com/jlnav/dragonfly.git@fix/remove_npobject
 
-
-        - name: Copy heffte exe on cache-hit
-          if: matrix.os == 'ubuntu-latest'
-          run: |
-            cd /home/runner/work/libensemble/libensemble
-            cp ./heffte/build/benchmarks/speed3d_c2c ./libensemble/tests/regression_tests/
-
         - name: Install other testing dependencies
           run: |
             conda install octave
