### 2023-12-08 62dde20
Don't build heffte for macOS
