### 2023-12-09 9f51157
adjustment to libE_specs["workerID"] default value in specs broke tcp. fixing due to libE logic (is_worker = libE_specs.get("workerID") is not None)
