### 2023-12-11 4aa88d5
QCommThread and QCommProcess now inherit from a QCommLocal
