### 2023-12-11 #6031 @jbytecode
@jmlarson1 - I've just sent a pull request. This PR includes a minor addition about adding a citation to referenced software. Please review the PR. If you agree with it, please merge. Thank you in advance.

https://github.com/Libensemble/libensemble/pull/1192
