### 2023-12-11 #6031 @jbytecode
@jmlarson1 - the paper and the bibtex seem better now. Thank you for merging my PR.

Now, please create a tagged release of the software in the repository. Using this tagged release, create a Zenodo archive. The archive should match the paper title, the author list, and their ORCIDs.

After performing those tasks, please report me the tagged release, the archive url, and the Zenodo DOI here.

Thank you in advance.
