### 2023-12-11 f01e834
removing a redundant QCommLocal.terminate(). within kill_proc_team, shouldn't need to call terminate on exception since .result already calls terminate
