### 2023-12-11 4086db7
undo changes to libE, looks like we always raised a timeout then needed to join
