+++ libensemble/alloc_funcs/start_only_persistent.py
@@ -107,7 +107,7 @@ def only_persistent_gens(W, H, sim_specs, gen_specs, alloc_specs, persis_info, l
 
     # Now the give_sim_work_first part
     points_to_evaluate = ~H["sim_started"] & ~H["cancel_requested"]
-    if alloc_specs["user"]["alt_type"]:
+    if user.get("alt_type", False):
         avail_workers = list(
             set(support.avail_worker_ids(persistent=False, zero_resource_workers=False))
             | set(support.avail_worker_ids(persistent=EVAL_SIM_TAG, zero_resource_workers=False))
@@ -122,12 +122,7 @@ def only_persistent_gens(W, H, sim_specs, gen_specs, alloc_specs, persis_info, l
 
         try:
             Work[wid] = support.sim_work(
-                wid,
-                H,
-                sim_specs["in"],
-                sim_ids_to_send,
-                persis_info.get(wid),
-                persistent=alloc_specs["user"].get("alt_type", False),
+                wid, H, sim_specs["in"], sim_ids_to_send, persis_info.get(wid), persistent=user.get("alt_type", False)
             )
         except InsufficientFreeResources:
             break
