+++ libensemble/alloc_funcs/start_only_persistent.py
@@ -6,32 +6,17 @@ from libensemble.tools.alloc_support import AllocSupport, InsufficientFreeResour
 
 def only_persistent_gens(W, H, sim_specs, gen_specs, alloc_specs, persis_info, libE_info):
     """
-    By default
-        This allocation function will give simulation work if possible, but
-        otherwise start up to ``alloc_specs["user"]["num_active_gens"]``
-        persistent generators (defaulting to one).
-
-        If any workers are marked as zero_resource_workers, then these will only
-        be used for generators.
-
-    If alloc_specs["user"]["alt_type"]:
-
-        This allocation function will give simulation work if possible to any worker
-        not listed as a zero_resource_worker. On the first call, the worker will be
-        placed into a persistent state that will be maintained until libE is exited.
-
-        Otherwise, zero resource workers will be given up to a maximum of
-        ``alloc_specs["user"]["num_active_gens"]`` persistent generators (defaulting to one).
-
-        Note, that an alternative to using zero resource workers would be to set
-        a fixed number of simulation workers in persistent state at the start, allowing
-        at least one worker for the generator - a minor alteration.
-
+    This allocation function will give simulation work if possible, but
+    otherwise start up to ``alloc_specs["user"]["num_active_gens"]``
+    persistent generators (defaulting to one).
 
     By default, evaluation results are given back to the generator once
     all generated points have been returned from the simulation evaluation.
     If ``alloc_specs["user"]["async_return"]`` is set to True, then any
     returned points are given back to the generator.
+
+    If any workers are marked as zero_resource_workers, then these will only
+    be used for generators.
 
     If any of the persistent generators has exited, then ensemble shutdown
     is triggered.
@@ -61,9 +46,6 @@ def only_persistent_gens(W, H, sim_specs, gen_specs, alloc_specs, persis_info, l
         `test_persistent_uniform_sampling_async.py <https://github.com/Libensemble/libensemble/blob/develop/libensemble/tests/functionality_tests/test_persistent_uniform_sampling_async.py>`_ # noqa
         `test_persistent_surmise_calib.py <https://github.com/Libensemble/libensemble/blob/develop/libensemble/tests/regression_tests/test_persistent_surmise_calib.py>`_ # noqa
         `test_persistent_uniform_gen_decides_stop.py <https://github.com/Libensemble/libensemble/blob/develop/libensemble/tests/functionality_tests/test_persistent_uniform_gen_decides_stop.py>`_ # noqa
-
-    For alloc_specs["user"]["alt_type"]:
-        `test_persistent_gensim_uniform_sampling.py <https://github.com/Libensemble/libensemble/blob/develop/libensemble/tests/functionality_tests/test_persistent_sim_uniform_sampling.py>`_ # noqa
     """
 
     if libE_info["sim_max_given"] or not libE_info["any_idle_workers"]:
@@ -107,13 +89,7 @@ def only_persistent_gens(W, H, sim_specs, gen_specs, alloc_specs, persis_info, l
 
     # Now the give_sim_work_first part
     points_to_evaluate = ~H["sim_started"] & ~H["cancel_requested"]
-    if user.get("alt_type", False):
-        avail_workers = list(
-            set(support.avail_worker_ids(persistent=False, zero_resource_workers=False))
-            | set(support.avail_worker_ids(persistent=EVAL_SIM_TAG, zero_resource_workers=False))
-        )
-    else:
-        avail_workers = support.avail_worker_ids(persistent=False, zero_resource_workers=False)
+    avail_workers = support.avail_worker_ids(persistent=False, zero_resource_workers=False)
     for wid in avail_workers:
         if not np.any(points_to_evaluate):
             break
[…]
