### 2023-12-12 0571611
Trying to support alt_type
