+++ libensemble/alloc_funcs/start_only_persistent.py
@@ -136,130 +136,3 @@ def only_persistent_gens(W, H, sim_specs, gen_specs, alloc_specs, persis_info, l
                 gen_count += 1
 
     return Work, persis_info, 0
-
-
-def only_persistent_workers(W, H, sim_specs, gen_specs, alloc_specs, persis_info, libE_info):
-    """
-    This allocation function will give simulation work if possible to any worker
-    not listed as a zero_resource_worker. On the first call, the worker will be
-    placed into a persistent state that will be maintained until libE is exited.
-
-    Otherwise, zero resource workers will be given up to a maximum of
-    ``alloc_specs["user"]["num_active_gens"]`` persistent generators (defaulting to one).
-
-    By default, evaluation results are given back to the generator once
-    all generated points have been returned from the simulation evaluation.
-    If ``alloc_specs["user"]["async_return"]`` is set to True, then any
-    returned points are given back to the generator.
-
-    If any of the persistent generators has exited, then ensemble shutdown
-    is triggered.
-
-    Note, that an alternative to using zero resource workers would be to set
-    a fixed number of simulation workers in persistent state at the start, allowing
-    at least one worker for the generator - a minor alteration.
-
-    **User options**:
-
-    To be provided in calling script: E.g., ``alloc_specs["user"]["async_return"] = True``
-
-    init_sample_size: int, optional
-        Initial sample size - always return in batch. Default: 0
-
-    num_active_gens: int, optional
-        Maximum number of persistent generators to start. Default: 1
-
-    async_return: Boolean, optional
-        Return results to gen as they come in (after sample). Default: False (batch return).
-
-    active_recv_gen: Boolean, optional
-        Create gen in active receive mode. If True, the manager does not need to wait
-        for a return from the generator before sending further returned points.
-        Default: False
-
-
-    .. seealso::
-        `test_persistent_gensim_uniform_sampling.py <https://github.com/Libensemble/libensemble/blob/develop/libensemble/tests/functionality_tests/test_persistent_sim_uniform_sampling.py>`_ # noqa
-    """
-
-    if libE_info["sim_max_given"] or not libE_info["any_idle_workers"]:
-        return {}, persis_info
-
-    # Initialize alloc_specs["user"] as user.
-    user = alloc_specs.get("user", {})
-    manage_resources = libE_info["use_resource_sets"]
-    active_recv_gen = user.get("active_recv_gen", False)  # Persistent gen can handle irregular communications
-    init_sample_size = user.get("init_sample_size", 0)  # Always batch return until this many evals complete
-    batch_give = user.get("give_all_with_same_priority", False)
-
-    support = AllocSupport(W, manage_resources, persis_info, libE_info)
-    gen_count = support.count_persis_gens()
-    Work = {}
-
-    # Asynchronous return to generator
-    async_return = user.get("async_return", False) and sum(H["sim_ended"]) >= init_sample_size
-
-    if gen_count < persis_info.get("num_gens_started", 0):
-        # When a persistent gen worker is done, trigger a shutdown (returning exit condition of 1)
-        return Work, persis_info, 1
-
-    # Give evaluated results back to a running persistent gen
-    for wid in support.avail_worker_ids(persistent=EVAL_GEN_TAG, active_recv=active_recv_gen):
-        gen_inds = H["gen_worker"] == wid
-        returned_but_not_given = np.logical_and.reduce((H["sim_ended"], ~H["gen_informed"], gen_inds))
-        if np.any(returned_but_not_given):
-            if async_return or support.all_sim_ended(H, gen_inds):
-                point_ids = np.where(returned_but_not_given)[0]
-                Work[wid] = support.gen_work(
-                    wid,
-                    gen_specs["persis_in"],
-                    point_ids,
[…]
