### 2023-12-13 bceaf55
Simplifying sine_tutorial wording
