### 2023-12-13 8e8daad
Correct Aurora/Sunspot GPU platform setting
