### 2023-12-13 2abb656
Add example Aurora submission script for forces_gpu
