### 2023-12-13 c5e2b0e
make some attributes of QCommThread and QCommProcess part of superclass
