### 2023-12-13 b539d71
re-add / implement attribute-setting validators for pydantic 1/2 models
