+++ libensemble/utils/validators.py
@@ -136,6 +136,34 @@ if pydanticV1:
 
         return values
 
+    @root_validator
+    def simf_set_in_out_from_attrs(cls, values):
+        if not values.get("sim_f"):
+            from libensemble.sim_funcs.one_d_func import one_d_example
+
+            values["sim_f"] = one_d_example
+        if hasattr(values.get("sim_f"), "inputs") and not values.get("inputs"):
+            values["inputs"] = values.get("sim_f").inputs
+        if hasattr(values.get("sim_f"), "outputs") and not values.get("outputs"):
+            values["out"] = values.get("sim_f").outputs
+        if hasattr(values.get("sim_f"), "persis_in") and not values.get("persis_in"):
+            values["persis_in"] = values.get("sim_f").persis_in
+        return values
+
+    @root_validator
+    def genf_set_in_out_from_attrs(cls, values):
+        if not values.get("gen_f"):
+            from libensemble.gen_funcs.sampling import latin_hypercube_sample
+
+            values["gen_f"] = latin_hypercube_sample
+        if hasattr(values.get("gen_f"), "inputs") and not values.get("inputs"):
+            values["inputs"] = values.get("gen_f").inputs
+        if hasattr(values.get("gen_f"), "outputs") and not values.get("outputs"):
+            values["out"] = values.get("gen_f").outputs
+        if hasattr(values.get("gen_f"), "persis_in") and not values.get("persis_in"):
+            values["persis_in"] = values.get("gen_f").persis_in
+        return values
+
     # RESOURCES VALIDATORS #####
 
     @root_validator
@@ -195,6 +223,26 @@ elif pydanticV2:
             assert hasattr(self.gen_specs, "gen_f"), "Generator function not provided to GenSpecs."
             assert isinstance(self.gen_specs.gen_f, Callable), "Generator function is not callable."
 
+        return self
+
+    @model_validator(mode="after")
+    def simf_set_in_out_from_attrs(self):
+        if hasattr(self.__dict__.get("sim_f"), "inputs") and not self.__dict__.get("inputs"):
+            self.__dict__["inputs"] = self.__dict__.get("sim_f").inputs
+        if hasattr(self.__dict__.get("sim_f"), "outputs") and not self.__dict__.get("outputs"):
+            self.__dict__["out"] = self.__dict__.get("sim_f").outputs
+        if hasattr(self.__dict__.get("sim_f"), "persis_in") and not self.__dict__.get("persis_in"):
+            self.__dict__["persis_in"] = self.__dict__.get("sim_f").persis_in
+        return self
+
+    @model_validator(mode="after")
+    def genf_set_in_out_from_attrs(self):
+        if hasattr(self.__dict__.get("gen_f"), "inputs") and not self.__dict__.get("inputs"):
+            self.__dict__["inputs"] = self.__dict__.get("gen_f").inputs
+        if hasattr(self.__dict__.get("gen_f"), "outputs") and not self.__dict__.get("outputs"):
+            self.__dict__["out"] = self.__dict__.get("gen_f").outputs
+        if hasattr(self.__dict__.get("gen_f"), "persis_in") and not self.__dict__.get("persis_in"):
+            self.__dict__["persis_in"] = self.__dict__.get("gen_f").persis_in
         return self
 
     # RESOURCES VALIDATORS #####
