+++ .github/workflows/extra.yml
@@ -228,4 +228,4 @@ jobs:
         runs-on: ubuntu-latest
         steps:
         - uses: actions/checkout@v4
-        - uses: crate-ci/typos@v1.16.24
+        - uses: crate-ci/typos@v1.16.25
