### 2023-12-13 8ff81ee
Testing/threading plus coverage (#1188)
* implement local threading into parse_argso

* add local_threading to 1d_sampling and run-tests

* specify additional threading concurrency to coverage line

* updating tests/.coveragerc

* rename "local_threading" to "threads"

* ensure we do an exception test with abort_on_exception being true

* adjustment to libE_specs["workerID"] default value in specs broke tcp. fixing due to libE logic (is_worker = libE_specs.get("workerID") is not None)

* fix validator

* add matplotlib to testing_requirements.txt

* that aint the right way

* replacing example
[…]
