+++ libensemble/libE.py
@@ -244,7 +244,7 @@ def libE(
         logger.manager_warning("Dry run. All libE() inputs validated. Exiting.")
         sys.exit()
 
-    libE_funcs = {"mpi": libE_mpi, "tcp": libE_tcp, "local": libE_local, "local_threading": libE_local}
+    libE_funcs = {"mpi": libE_mpi, "tcp": libE_tcp, "local": libE_local, "threads": libE_local}
 
     Resources.init_resources(libE_specs, platform_info)
     if Executor.executor is not None:
@@ -439,7 +439,7 @@ def start_proc_team(nworkers, sim_specs, gen_specs, libE_specs, log_comm=True):
 
     if libE_specs["comms"] == "local":
         QCommLocal = QCommProcess
-    else:  # local_threading
+    else:  # threads
         QCommLocal = QCommThread
         log_comm = False  # Prevents infinite loop of logging.
 
