+++ examples/tutorials/simple_sine/test_local_sine_tutorial_2.py
@@ -1,0 +1,1 @@
+../../../libensemble/tests/functionality_tests/test_local_sine_tutorial_2.py
