### 2023-12-14 4f7cd89
Bump globus-compute-sdk from 2.8.0 to 2.9.0
Bumps [globus-compute-sdk](https://github.com/funcx-faas/funcx) from 2.8.0 to 2.9.0.
- [Release notes](https://github.com/funcx-faas/funcx/releases)
- [Changelog](https://github.com/funcx-faas/funcX/blob/main/docs/changelog.rst)
- [Commits](https://github.com/funcx-faas/funcx/commits)

---
updated-dependencies:
- dependency-name: globus-compute-sdk
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
