+++ install/misc_feature_requirements.txt
@@ -1,1 +1,1 @@
-globus-compute-sdk==2.8.0
+globus-compute-sdk==2.9.0
