### 2023-12-14 32e196c
Feature / user-function input/output field decorators (#1072)
* tentatively-named decorators for setting user-function data attributes. SimSpecs and GenSpecs can extract and set

* wrappers append input/output attributes to top of wrapped function docstrings

* bold labels

* add persistent_input_fields decorator

* initial docstrings for decorator functions

* first pass on briefly mentioning decorators in generator function guide

* use synced tabs to compare decorated and non-decorated functions and their specs

* add import line for decorators, fix specs formatting

* re-add / implement at
[…]
