### 2023-12-14 5270422
fix check_exit_criteria validator bug, add separate pydantic1 jobs to extra-CI's matrix
