+++ .github/workflows/extra.yml
@@ -12,15 +12,18 @@ jobs:
                 os: [ubuntu-latest]
                 mpi-version: [mpich]
                 python-version: [3.9, "3.10", "3.11", "3.12"]
+                pydantic-version: ["2.5.2"]
                 comms-type: [m, l]
                 include:
                     - os: macos-latest
                       python-version: 3.11
                       mpi-version: "mpich=4.0.3"
+                      pydantic-version: "2.5.2"
                       comms-type: m
                     - os: macos-latest
                       python-version: 3.11
                       mpi-version: "mpich=4.0.3"
+                      pydantic-version: "2.5.2"
                       comms-type: l
                     - os: ubuntu-latest
                       python-version: "3.10"
@@ -29,6 +32,16 @@ jobs:
                     - os: ubuntu-latest
                       mpi-version: "openmpi"
                       python-version: "3.12"
+                      comms-type: l
+                    - os: ubuntu-latest
+                      mpi-version: mpich
+                      python-version: "3.10"
+                      pydantic-version: "1.10.13"
+                      comms-type: m
+                    - os: ubuntu-latest
+                      mpi-version: mpich
+                      python-version: "3.10"
+                      pydantic-version: "1.10.13"
                       comms-type: l
 
         env:
@@ -64,7 +77,7 @@ jobs:
               /usr/share/miniconda3/bin
               /usr/share/miniconda3/lib
               /usr/share/miniconda3/include
-            key: libe-${{ github.ref_name }}-${{ matrix.python-version }}-${{ matrix.comms-type }}-extra
+            key: libe-${{ github.ref_name }}-${{ matrix.python-version }}-${{ matrix.comms-type }}-${{ matrix.pydantic-version }}-extra
 
         - name: Force-update certifi
           run: |
@@ -190,6 +203,7 @@ jobs:
 
         - name: Install libEnsemble, flake8, lock environment
           run: |
+            pip install pydantic==${{ matrix.pydantic-version }}
             pip install -e .
             flake8 libensemble
 
