### 2023-12-15 da2dcca
Support multi-tile GPUs if in platform specs
