+++ libensemble/resources/resources.py
@@ -9,6 +9,7 @@ import socket
 
 from libensemble.resources import node_resources
 from libensemble.resources.env_resources import EnvResources
+from libensemble.resources.gpu_detect import get_num_tiles
 from libensemble.resources.mpi_resources import get_MPI_runner
 from libensemble.resources.worker_resources import ResourceManager, WorkerResources
 
@@ -178,7 +179,7 @@ class GlobalResources:
         self.tiles_per_gpu = 1
         if libE_specs.get("use_tiles_as_gpus", False):
             # Not yet detected so only uses tiles if set in platform_info
-            self.tiles_per_gpu = platform_info.get("tiles_per_gpu") or 1
+            self.tiles_per_gpu = platform_info.get("tiles_per_gpu") or get_num_tiles()
 
         node_file = resource_info.get("node_file", None)
         nodelist_env_slurm = resource_info.get("nodelist_env_slurm", None)
