### 2023-12-15 ca86d2b
Show tile examples in submission script
