### 2023-12-15 d05dc3a
remove globus-compute and balsam tests if running with pydantic 2
