+++ .github/workflows/extra.yml
@@ -215,6 +215,12 @@ jobs:
             rm ./libensemble/tests/regression_tests/test_persistent_gp.py
             rm ./libensemble/tests/regression_tests/test_persistent_gp_multitask_ax.py
 
+        - name: Remove Balsam/Globus-compute tests on Pydantic 2
+          if: matrix.pydantic-version == '2.5.2'
+          run: |
+            rm ./libensemble/tests/unit_tests/test_ufunc_runners.py
+            rm ./libensemble/tests/unit_tests/test_executor_balsam.py
+
         - name: Run extensive tests, Ubuntu
           if: matrix.os == 'ubuntu-latest'
           run: |
