### 2023-12-15 8124e65
Add Aurora user guide
