### 2023-12-15 fad9c22
Add path to Aurora submission script in guide
