### 2023-12-15 de507fa
add missing pydantic versions to openmpi and tcp jobs
