+++ .github/workflows/extra.yml
@@ -28,9 +28,11 @@ jobs:
                     - os: ubuntu-latest
                       python-version: "3.10"
                       mpi-version: "mpich"
+                      pydantic-version: "2.5.2"
                       comms-type: t
                     - os: ubuntu-latest
                       mpi-version: "openmpi"
+                      pydantic-version: "2.5.2"
                       python-version: "3.12"
                       comms-type: l
                     - os: ubuntu-latest
