### 2023-12-15 91c82d4
Put back gpu_setting_type
