### 2023-12-15 2361a4b
Fix test_platform.py
