### 2023-12-15 eccad2d
Remove unnecessary float conversion
