### 2023-12-15 b0381a3
Fix a validator bug, add pydantic2 to extra-tests matrix (#1197)
* fix check_exit_criteria validator bug, add separate pydantic1 jobs to extra-CI's matrix

* forgot to re-add alias for alloc_specs out

* remove globus-compute and balsam tests if running with pydantic 2

* add missing pydantic versions to openmpi and tcp jobs

* AttributeError can be raised from attempting to import BalsamExecutor with Pydantic2, so lets also catch that to avoid importing the BalsamExecutor
