+++ libensemble/tests/regression_tests/test_1d_sampling.py
@@ -30,6 +30,7 @@ if __name__ == "__main__":
     sampling.sim_specs = SimSpecs(sim_f=sim_f)
     sampling.gen_specs = GenSpecs(
         gen_f=gen_f,
+        outputs=[("x", float, (1,))],
         user={
             "gen_batch_size": 500,
             "lb": np.array([-3]),
