+++ libensemble/tests/regression_tests/test_persistent_aposmm_scipy.py
@@ -50,6 +50,8 @@ if __name__ == "__main__":
     n = 2
     sim_specs = {
         "sim_f": sim_f,
+        "in": ["x"],
+        "out": [("f", float)],
     }
 
     gen_out = [
