+++ libensemble/tests/regression_tests/test_persistent_tasmanian.py
@@ -64,6 +64,8 @@ if __name__ == "__main__":
     num_dimensions = 2
     sim_specs = {
         "sim_f": sim_f,
+        "in": ["x"],
+        "out": [("f", float)],
     }
 
     gen_specs = {
