### 2023-12-17 #6031 @jbytecode
@jmlarson1 - we have a rendering problem. The part

```
Additional examples of applications that have used libEnsemble are
surveyed in [Representative libEnsemble Use Cases](#Representative-libEnsemble-Use-Cases).
```

is the root of the problem. Could you please consider correcting it using a link maybe rather than a footnote?
