+++ docs/papers/joss/paper.bib
@@ -80,7 +80,7 @@
 
 @article{PhysRevAccelBeams.26.084601,
   title = {Bayesian optimization of laser-plasma accelerators assisted by reduced physical models},
-  author = {Ferran Pousa, A. and Jalas, S. and Kirchen, M. and Martinez de la Ossa, A. and Th\'evenet, M. and Hudson, S. and Larson, J. and Huebl, A. and Vay, J.-L. and Lehe, R.},
+  author = {{Ferran Pousa}, A. and Jalas, S. and Kirchen, M. and Martinez de la Ossa, A. and Th\'evenet, M. and Hudson, S. and Larson, J. and Huebl, A. and Vay, J.-L. and Lehe, R.},
 	journal = {Physical Review Accelerators and Beams},
   volume = {26},
   issue = {8},
@@ -95,7 +95,7 @@
 
 @article{Pousa22,
 year = {2022},
-    author = {Ferran Pousa, A and Jalas, S. and Kirchen, M. and Martinez de la Ossa, A. and Thévenet, M. and Hudson, S. and Larson, J.  and Huebl, A. and Vay, J.-L. and Lehe, R.},
+    author = {{Ferran Pousa}, A and Jalas, S. and Kirchen, M. and Martinez de la Ossa, A. and Thévenet, M. and Hudson, S. and Larson, J.  and Huebl, A. and Vay, J.-L. and Lehe, R.},
     title = {Multitask Optimization of Laser-Plasma Accelerators Using Simulation Codes with Different Fidelities},
     pages = {1761--1764},
     paper = {WEPOST030},
