### 2023-12-18 #6031 @jmlarson1
@jbytecode The anchor/crossref to the section header wasn't working correctly, so I've explicitly added a markdown anchor.

I find that pointing to that has resolved the issue, right?
