### 2023-12-18 #6031 @danielskatz
@jmlarson1 - I've suggested a few small changes in https://github.com/Libensemble/libensemble/pull/1199 - please merge this, or let me know what you disagree with, and then we can proceed
