### 2023-12-18 #6031 @danielskatz
Congratulations to @jmlarson1 (Jeffrey Larson) and co-authors on your publication!!

JOSS completely depends on volunteers, and we couldn't be successful without you
